==> 03-community_composition.Rmd <==
# Community composition

```{r load_data03, include=TRUE}
rm(list=ls()) #clear environment
load("data/squirrels_data.Rdata")
```

## Count data preparation

```{r data_preparation03, warning=FALSE}

genome_counts_log <- genome_counts %>% 
    column_to_rownames(var="genome") %>%
    mutate_all(~log10(.+1)) #fixed: mutate_at(vars(), ~log10(.+1))) was not working

genome_counts_pivot <- genome_counts %>%
  mutate_at(vars(-genome),~./sum(.)) %>% #apply TSS normalisation
  pivot_longer(-genome, names_to = "sample", values_to = "count") %>% #reduce to minimum number of columns
  left_join(., genome_metadata, by = join_by(genome == genome)) 
# %>% #append taxonomy
  # mutate(phylum = fct_relevel(phylum, rev(ehi_phylum_colors$phylum))) #sort phyla by taxonomy

genome_counts_by_host <- sample_metadata %>%
  select("sample","species","area_type", "development", "season") %>%
  rename(host_sp=species) %>%
  left_join(genome_counts_pivot,., by=join_by("sample" == "sample")) #%>%

  
# Which host species each genome can be found in
genomes_by_species <- genome_counts_by_host %>%
  filter(count>0) %>%
  group_by(genome) %>%
  mutate(host = if_else(all(host_sp == "Sciurus vulgaris"), "only red",
                        if_else(all(host_sp == "Sciurus carolinensis"), "only grey", "both"))) %>%
  select(genome, host) %>%
  distinct(genome, .keep_all = TRUE) %>%
  left_join(.,genome_metadata, by='genome')

genomes_by_species$host <-factor(genomes_by_species$host, levels = c("both", "only red", "only grey"))

genomes_by_species %>%
  group_by(host) %>%
  summarise(n=length(completeness),
            m_compl=mean(completeness),
            sd_compl=sd(completeness),
            m_cont=mean(contamination),
            sd_cont=sd(contamination))

compl_red <- genomes_by_species %>%
  filter(host %in% c("only red", "both")) %>%
  pull(completeness)
compl_grey <- genomes_by_species %>%
  filter(host %in% c("only grey", "both")) %>%
  pull(completeness)

mean(compl_red)
sd(compl_red)
mean(compl_grey)
sd(compl_grey)

wilcox.test(compl_red, compl_grey)

cont_red <- genomes_by_species %>%
  filter(host %in% c("only red", "both")) %>%
  pull(contamination)
cont_grey <- genomes_by_species %>%
  filter(host %in% c("only grey", "both")) %>%
  pull(contamination)

mean(cont_red)
sd(cont_red)

mean(cont_grey)
sd(cont_grey)

wilcox.test(cont_red, cont_grey)

a <- genomes_by_species %>%
  filter(host %in% c("only red", "both"))
b <- genomes_by_species %>%
  filter(host %in% c("only grey", "both"))
c <- genomes_by_species %>%
  filter(host %in% c("only red", "only grey")) 

wilcox.test(completeness ~ host, data = a)
wilcox.test(completeness ~ host, data = b)
wilcox.test(completeness ~ host, data = c)
  
wilcox.test(contamination ~ host, data = a)
wilcox.test(contamination ~ host, data = b)
wilcox.test(contamination ~ host, data = c)


```

## Overview of genomes by host species

```{r mags_by_species_tree, fig.height=8, fig.width=10, fig.fullwidth=TRUE, warning=FALSE, message=FALSE}

# Generate the phylum color heatmap
phylum_heatmap <- genome_metadata %>%
  arrange(match(genome, genome_tree$tip.label)) %>%
  select(genome,phylum) %>%
  mutate(phylum = factor(phylum, levels = unique(phylum))) %>%
  column_to_rownames(var = "genome")

# Create baseline circular genome tree
circular_tree <- force.ultrametric(genome_tree,method="extend") %>%
  ggtree(., layout = 'circular', size = 0.1, angle=45) +
  xlim(-1, NA)

# Add phylum colors ring
circular_tree <- gheatmap(circular_tree, phylum_heatmap, offset=0.0, width=0.3, colnames=FALSE) +
  scale_fill_manual(values=custom_colors, name="Phylum") +
  #geom_tiplab2(size=1, hjust=-0.1) +
  theme(legend.position = "right", plot.margin = margin(0, 0, 0, 0), panel.margin = margin(0, 0, 0, 0))


#Flush color scale to enable a new color scheme in the next ring
circular_tree <- circular_tree + new_scale_fill()

# Add host ring
circular_tree <-  circular_tree +
  new_scale_fill() +
  scale_fill_manual(values = c("black", "#ed2939", "#92a0ad"), name="Host\nspecies") + #"#cc3333", "#999999"
  geom_fruit(
    data=genomes_by_species,
    geom=geom_tile,
    mapping = aes(y=genome, fill=host),
    offset = 0.55,
    width=0.3)
  

#Plot circular tree
circular_tree


```

```{r venn, message=FALSE, warning=FALSE}


#MAGs by host species
genomes_by_species %>%
  dplyr::group_by(host) %>%
  summarise(n=length(host),
            percentage=(length(host)/1687)*100) %>%
  tt()

genomes_by_species %>%
  mutate(Sc = host %in% c("only grey", "both"),
        Sv = host %in% c("only red", "both")) %>%
  select(genome,Sc,Sv) %>%
  ggplot() +
  geom_venn(aes(A = `Sv`, B = `Sc`), , fill_color = c("#cc3333", "#999999"), fill_alpha=0.5,
            stroke_color = c("#cc3333", "#999999"), stroke_size=0.7) +
  coord_fixed() +
  theme_void()

# mags by urbanisation in red
venn_urb_red <- genome_counts_by_host %>%
  filter(host_sp == "Sciurus vulgaris",
         count>0) %>%
  group_by(genome) %>%
  summarise(
    urban = any(area_type == "urban"),
    suburban = any(area_type == "suburban"),
    rural = any(area_type == "rural")) %>%
  select(genome,urban, suburban, rural) %>%
  ggplot() +
  geom_venn(aes(A = `urban`, B = `suburban`, C = `rural`), fill_color = c("#c1cca7", "#94a766", "#3d6a4b"), fill_alpha=0.5,
            stroke_color = c("#c1cca7", "#94a766", "#3d6a4b"), stroke_size=0.7, set_name_size = 5) +
  coord_fixed() +
  theme_void()

# mags by urbanisation in grey
venn_urb_grey <- genome_counts_by_host %>%
  filter(host_sp == "Sciurus carolinensis",
         count>0) %>%
  group_by(genome) %>%
  summarise(
    urban = any(area_type == "urban"),
    suburban = any(area_type == "suburban"),
    rural = any(area_type == "rural")) %>%
  select(genome,urban, suburban, rural) %>%
  ggplot() +
  geom_venn(aes(A = `urban`, B = `suburban`, C = `rural`), , fill_color = c("#c1cca7", "#94a766", "#3d6a4b"), fill_alpha=0.5,
            stroke_color = c("#c1cca7", "#94a766", "#3d6a4b"), stroke_size=0.7, set_name_size = 5) +
  coord_fixed() +
  theme_void()

ggarrange(venn_urb_red, venn_urb_grey, nrow=1, ncol=2, labels =c("Sv", "Sc"))

# mags by season in red
venn_seas_red <- genome_counts_by_host %>%
  filter(host_sp == "Sciurus vulgaris",
         count>0) %>%
  group_by(genome) %>%
  summarise(
    spring = any(season == "spring-summer"),
    autumn = any(season == "autumn"),
    winter = any(season == "winter")) %>%
  select(genome, spring, autumn, winter) %>%
  ggplot() +
  geom_venn(aes(A = `spring`, B = `autumn`, C = `winter`), , fill_color = c("#76b183", "#d98a41", "#6b7398"), fill_alpha=0.5,
            stroke_color = c("#76b183", "#d98a41", "#6b7398"), stroke_size=0.7, set_name_size = 5) +
  coord_fixed() +
  theme_void()

# mags by season in grey
venn_seas_grey <- genome_counts_by_host %>%
  filter(host_sp == "Sciurus carolinensis",
         count>0) %>%
  group_by(genome) %>%
  summarise(
    spring = any(season == "spring-summer"),
    autumn = any(season == "autumn"),
    winter = any(season == "winter")) %>%
  select(genome, spring, autumn, winter) %>%
  ggplot() +
  geom_venn(aes(A = `spring`, B = `autumn`, C = `winter`), , fill_color = c("#76b183", "#d98a41", "#6b7398"), fill_alpha=0.5,
            stroke_color = c("#76b183", "#d98a41", "#6b7398"), stroke_size=0.7, set_name_size = 5) +
  coord_fixed() +
  theme_void()

ggarrange(venn_seas_red, venn_seas_grey, nrow=1, ncol=2, labels =c("Sv", "Sc"))

```



## Genomes counts


```{r mags_counts, fig.width=10, fig.height=8, fig.fullwidth=TRUE, warning=FALSE, message=FALSE}

vertical_tree <- force.ultrametric(genome_tree,method="extend") %>%
  ggtree(., size = 0.3)

#Add phylum colors
vertical_tree <- gheatmap(vertical_tree, phylum_heatmap, offset=0, width=0.1, colnames=FALSE, color=NA) +
  scale_fill_manual(values=custom_colors)

#Reset fill scale
vertical_tree <- vertical_tree + new_scale_fill()

#Add counts
vertical_tree <- gheatmap(vertical_tree, genome_counts_log, offset=0.5, width=3.5, color=NA, colnames=FALSE) + #, colnames_angle=90, font.size=2, colnames_position="top", colnames_offset_y = 9
  vexpand(.08) +
  coord_cartesian(clip = "off") +
  scale_fill_gradient(low = "#f4f4f4", high = "steelblue", na.value="white")

#Plot tree
vertical_tree +
  theme(legend.position='right')
```

```{r mean_ab_shared}

mags_both <- genomes_by_species %>%
  filter(host=="both") %>%
  pull(genome)
  
mags_red <- genomes_by_species %>%
  filter(host=="only red") %>%
  pull(genome)

mags_grey <- genomes_by_species %>%
  filter(host=="only grey") %>%
  pull(genome)

genome_counts_by_host %>%
  filter(genome %in% mags_both) %>%
  group_by(sample) %>%
  summarise(sum_abundance=sum(count)) %>%
  ungroup() %>%
  summarise(mean_abundance=mean(sum_abundance),
            sd=sd(sum_abundance)) 

genome_counts_by_host %>%
  filter(genome %in% mags_red) %>%
  group_by(sample) %>%
  summarise(sum_abundance=sum(count)) %>%
  ungroup() %>%
  summarise(mean_abundance=mean(sum_abundance),
            sd=sd(sum_abundance)) 

genome_counts_by_host %>%
  filter(genome %in% mags_grey) %>%
  group_by(sample) %>%
  summarise(sum_abundance=sum(count)) %>%
  ungroup() %>%
  summarise(mean_abundance=mean(sum_abundance),
            sd=sd(sum_abundance)) 

```


## Taxonomic composition of samples


```{r taxonomic_composition_by_sample, fig.width=12, fig.height=8, fig.fullwidth=TRUE, warning=FALSE, message=FALSE}

# Plot stacked barplot
ggplot(genome_counts_by_host, aes(x=sample,y=count,fill=phylum, group=phylum))+ #grouping enables keeping the same sorting of taxonomic units
  geom_bar(stat="identity", colour="white", linewidth=0.02)+ #plot stacked bars with white borders
  scale_fill_manual(values=custom_colors, name="Phylum") +
  labs(y = "Relative abundance") +
  guides(fill = guide_legend(ncol = 1)) +
  facet_nested(~host_sp, scales="free", space="free") +
  theme(axis.text.x = element_blank(),
        axis.title.x = element_blank(),
        panel.background = element_blank(),
        panel.border = element_blank(),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        axis.line = element_line(linewidth = 0.5, linetype = "solid", colour = "black"),
        legend.position="right",
        )

#mean number of MAGs by individual
genome_counts_by_host %>%
  filter(count != 0) %>%                 
  group_by(sample) %>%                   
  summarise(n_genomes = n_distinct(genome)) %>% 
  summarise(avg_genomes = mean(n_genomes),
            sd_genomes = sd(n_genomes))
  
#red squirrels
genome_counts_by_host %>%
  filter(host_sp =="Sciurus vulgaris",
         count != 0) %>%                 
  group_by(sample) %>%                   
  summarise(n_genomes = n_distinct(genome)) %>% 
  summarise(mean_genomes = mean(n_genomes),
            sd_genomes = sd(n_genomes))

#grey squirrels
genome_counts_by_host %>%
  filter(host_sp =="Sciurus carolinensis",
         count != 0) %>%                 
  group_by(sample) %>%                   
  summarise(n_genomes = n_distinct(genome)) %>% 
  summarise(mean_genomes = mean(n_genomes),
            sd_genomes = sd(n_genomes))


```

## Taxonomic composition by host species 

### Phylum level

```{r phyla_by_species, fig.width=10, fig.height=8, fig.fullwidth=TRUE, message=FALSE, warning=FALSE}

prev_phyla_Sv <- genome_counts_by_host %>%
  filter(host_sp=="Sciurus vulgaris") %>%
  mutate(n = n_distinct(sample)) %>%
  ungroup() %>%
  filter(count>0) %>%
  group_by(phylum) %>%
  summarise(
    pos = n_distinct(sample),
    n = first(n),
    prevalence = (pos / n)*100
  ) 

abun_phyla_Sv <- genome_counts_by_host %>%
  filter(host_sp=="Sciurus vulgaris") %>%
  group_by(sample, phylum) %>%
  summarise(sum_abundance=sum(count)) %>%
  group_by(phylum) %>%
  summarise(mean_abundance=mean(sum_abundance),
            sd=sd(sum_abundance)) 

#prevalence and abundance of phyla in S. vulgaris
prev_phyla_Sv %>%
  left_join(abun_phyla_Sv, by = "phylum") %>%
  arrange(desc(prevalence)) %>%
  paged_table()

prev_phyla_Sc <- genome_counts_by_host %>%
  filter(host_sp=="Sciurus carolinensis") %>%
  mutate(n = n_distinct(sample)) %>%
  ungroup() %>%
  filter(count>0) %>%
  group_by(phylum) %>%
  summarise(
    pos = n_distinct(sample),
    n = first(n),
    prevalence = (pos / n)*100
  ) 

abun_phyla_Sc <- genome_counts_by_host %>%
  filter(host_sp=="Sciurus carolinensis") %>%
  group_by(sample, phylum) %>%
  summarise(sum_abundance=sum(count)) %>%
  group_by(phylum) %>%
  summarise(mean_abundance=mean(sum_abundance),
            sd=sd(sum_abundance)) 

#prevalence and abundance of phyla in S. carolinensis
prev_phyla_Sc %>%
  left_join(abun_phyla_Sc, by = "phylum") %>%
  arrange(desc(prevalence)) %>%
  paged_table()


phylum_summary <- genome_counts_by_host %>%
  group_by(sample,host_sp,phylum) %>%
  summarise(relabun=sum(count))

# phylum_summary %>%
#     group_by(phylum) %>%
#     summarise(mean=mean(relabun, na.rm=TRUE),sd=sd(relabun, na.rm=TRUE)) %>%
#     paged_table()

phylum_arrange <- phylum_summary %>%
    group_by(phylum) %>%
    summarise(mean=mean(relabun)) %>%
    arrange(-mean) %>%
    select(phylum) %>%
    pull()

phylum_summary %>%
    filter(phylum %in% phylum_arrange) %>%
    mutate(phylum=factor(phylum,levels=rev(phylum_arrange))) %>%
    ggplot(aes(x=relabun, y=phylum, group=phylum, color=phylum)) +
        scale_color_manual(values=custom_colors, name="Phylum") +
        geom_jitter(alpha=0.5) + 
        facet_nested(~host_sp, scales="free", space="free") +
        theme_minimal() + 
        theme(legend.position="none") +
        labs(y="Phylum",x="Relative abundance") 




```

### Family level

```{r families_by_species, fig.width=10, fig.height=8, fig.fullwidth=TRUE, message=FALSE, warning=FALSE}

#most prevalent families by host species
genome_counts_by_host %>%
  group_by(host_sp) %>%
  mutate(n = n_distinct(sample)) %>%
  ungroup() %>%
  filter(count>0) %>%
  group_by(host_sp, phylum, family) %>%
  summarise(
    pos = n_distinct(sample),
    n = first(n),
    prevalence = (pos / n)*100
  ) %>%
  group_by(host_sp) %>%
  top_n(5, prevalence) %>%
  arrange(host_sp, desc(prevalence)) %>%
  paged_table()

#most abundant families by host species
genome_counts_by_host %>%
  group_by(host_sp, sample, phylum, family) %>%
  summarise(sum_abundance = sum(count)) %>%
  group_by(host_sp, phylum, family) %>%
  summarise(mean_abundance=mean(sum_abundance),
            sd=sd(sum_abundance)) %>%
  arrange(host_sp, desc(mean_abundance)) %>%  # Sort within each host_sp group by abundance
  group_by(host_sp) %>%
  slice_max(order_by = mean_abundance, n = 5) %>%  # Select the top 5 families per host_sp
  ungroup() %>%
  arrange(host_sp, desc(mean_abundance)) %>%
  paged_table()


family_summary <- genome_counts_by_host %>%
  group_by(sample,host_sp,family) %>%
  summarise(relabun=sum(count))

# family_summary %>%
#     group_by(family) %>%
#     summarise(mean=mean(relabun, na.rm=TRUE),sd=sd(relabun, na.rm=TRUE)) %>%
#     arrange(-mean) %>%
#     paged_table()

family_arrange <- family_summary %>%
    group_by(family) %>%
    summarise(mean=sum(relabun)) %>%
    arrange(-mean) %>%
    select(family) %>%
    pull()

family_summary %>%
    left_join(genome_metadata %>% select(family,phylum) %>% unique(),by=join_by(family==family)) %>%
    left_join(sample_metadata,by=join_by(sample==sample)) %>%
    filter(family %in% family_arrange[1:20]) %>%
    mutate(family=factor(family,levels=rev(family_arrange[1:20]))) %>%
    filter(relabun > 0) %>%
    ggplot(aes(x=relabun, y=family, group=family, color=phylum)) +
        scale_color_manual(values=custom_colors) +
        geom_jitter(alpha=0.5) + 
        facet_grid(.~host_sp)+
        theme_minimal() + 
        labs(y="Family", x="Relative abundance", color="Phylum") + 
    guides(colour = guide_legend(override.aes = list(size=4, alpha=0.9)))

```



### Genus level

```{r genera_by_species, fig.height=8, fig.fullwidth=TRUE, message=FALSE, warning=FALSE}

#most prevalent genera by host species
genome_counts_by_host %>%
  group_by(host_sp) %>%
  mutate(N = n_distinct(sample)) %>%
  ungroup() %>%
  filter(count>0) %>%
  group_by(host_sp, phylum, family, genus) %>%
  summarise(
    pos = n_distinct(sample),
    N = first(N),
    prevalence = (pos / N)*100
  ) %>%
  group_by(host_sp) %>%
  top_n(5, prevalence) %>%
  arrange(host_sp, desc(prevalence)) %>%
  paged_table()

#most abundant genera by host species
genome_counts_by_host %>%
  group_by(host_sp, sample, phylum, family, genus) %>%
  summarise(sum_abundance = sum(count)) %>%
  group_by(host_sp, phylum, family, genus) %>%
  summarise(mean_abundance=mean(sum_abundance),
            sd=sd(sum_abundance)) %>%
  arrange(host_sp, desc(mean_abundance)) %>%  # Sort within each host_sp group by abundance
  group_by(host_sp) %>%
  slice_max(order_by = mean_abundance, n = 5) %>%  # Select the top 5 families per host_sp
  ungroup() %>%
  arrange(host_sp, desc(mean_abundance)) %>%
  paged_table()


#NB: there are several unnamed genera from different families that get grouped together

genus_summary <- genome_counts_by_host %>%
  group_by(sample,host_sp,genus) %>%
  summarise(relabun=sum(count))

# genus_summary %>%
#     group_by(genus) %>%
#     summarise(mean=mean(relabun, na.rm=TRUE),sd=sd(relabun, na.rm=TRUE)) %>%
#     arrange(-mean) %>%
#     paged_table()

genus_arrange <- genus_summary %>%
    group_by(genus) %>%
    summarise(mean=sum(relabun)) %>%
    arrange(-mean) %>%
    select(genus) %>%
    pull()

genus_summary %>%
    left_join(genome_metadata %>% select(genus,phylum) %>% unique(),by=join_by(genus==genus)) %>%
    left_join(sample_metadata,by=join_by(sample==sample)) %>%
    filter(genus %in% genus_arrange[1:20]) %>%
    mutate(genus=factor(genus,levels=rev(genus_arrange[1:20]))) %>%
    filter(relabun > 0) %>%
    ggplot(aes(x=relabun, y=genus, group=genus, color=phylum)) +
        scale_color_manual(values=custom_colors) +
        geom_jitter(alpha=0.5) + 
        facet_grid(.~host_sp)+
        theme_minimal() + 
        labs(y="Genus", x="Relative abundance", color="Phylum") + 
    guides(colour = guide_legend(override.aes = list(size=5, alpha=0.9)))

```


```{r core_genera, message=FALSE, warning=FALSE}

prevalence_gen <- genome_counts %>%
  pivot_longer(!genome, names_to = "sample", values_to = "counts") %>%
  left_join(sample_metadata %>% select(sample, species, metapop), by = "sample") %>%
  rename(host_sp = species) %>%
  left_join(genome_metadata, by = "genome") %>% 
  group_by(sample, family,genus) %>% # presence/absence in each sample
  summarise(genus_present = any(counts > 0), .groups = "drop") %>%
  left_join(sample_metadata %>% select(sample, species, metapop), by = "sample") %>%
  rename(host_sp = species) %>%
  group_by(family,genus, host_sp) %>% # group to summarize prevalence
  summarise(
    presence = sum(genus_present, na.rm = TRUE),
    sample_size = n_distinct(sample),
    metapop = n_distinct(metapop[genus_present]),
    prevalence = presence / sample_size,
    .groups = "drop"
  )

#core genera in red squirrel
prevalence_gen %>%
  filter(host_sp=="Sciurus vulgaris",
         prevalence > 0.75,
         metapop == 6) %>%
  arrange(-prevalence) %>%
  paged_table()

#core genera in grey squirrel
prevalence_gen %>%
  filter(host_sp=="Sciurus carolinensis",
         prevalence > 0.75,
         metapop == 4) %>%
  arrange(-prevalence) %>%
  paged_table()

```
