==> 04-Functional analysis.Rmd <==
---
title: "Functional analysis"
author: "Antton Alberdi"
date: "11/19/2023"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r GIFTs_data_preparation}
#install_github("anttonalberdi/distillR")
library(distillR)
library(ggplot2)
library(RColorBrewer)
library(reshape2)

tss <- function(abund){sweep(abund, 2, colSums(abund), FUN="/")}

GIFTs_elements <- distillr_table
genome_counts <- count_table_cov_size

#Aggregate element-level GIFTs into the function level
GIFTs_functions <- to.functions(GIFTs_elements,GIFT_db)

#Aggregate function-level GIFTs into overall Biosynthesis, Degradation and Structural GIFTs
GIFTs_domains <- to.domains(GIFTs_functions,GIFT_db)

#Get overall metabolic capacity indices per MAG (at the domain level)
rowMeans(GIFTs_functions) # averaged at the function level (each function is weighed equally)
rowMeans(GIFTs_domains) # averaged at the domain level (each domain is weighed equally)

#Get community-weighed average GIFTs per sample
# GIFTs_elements_community <- to.community(GIFTs_elements,genome_counts,GIFT_db)
# GIFTs_functions_community <- to.community(GIFTs_functions,genome_counts,GIFT_db)
# GIFTs_domains_community <- to.community(GIFTs_domains,genome_counts,GIFT_db)


GIFTs_elements_community <- count_table_cov_size %>%
  tss() %>%
  to.community(GIFTs_elements,.,GIFT_db)

GIFTs_functions_community <- count_table_cov_size %>%
  tss() %>%
  to.community(GIFTs_functions,.,GIFT_db)

GIFTs_domains_community <- count_table_cov_size %>%
  tss() %>%
  to.community(GIFTs_domains,.,GIFT_db)

GIFTs_elements_community %>%
  rowMeans() %>%
  as_tibble(., rownames = "sample") %>%
  left_join(sample_table, by = join_by(sample == sample))  %>%
  group_by(species) %>%
  summarise(MCI = mean(value), sd = sd(value))

GIFTs_functions_community %>%
  rowMeans() %>%
  as_tibble(., rownames = "sample") %>%
  left_join(sample_table, by = join_by(sample == sample))  %>%
  group_by(species) %>%
  summarise(MCI = mean(value), sd = sd(value))

GIFTs_domains_community %>%
  rowMeans() %>%
  as_tibble(., rownames = "sample") %>%
  left_join(sample_table, by = join_by(sample == sample))  %>%
  group_by(species) %>%
  summarise(MCI = mean(value), sd = sd(value))

merge_gift <- GIFTs_domains_community %>% 
  as.data.frame() %>% 
  rownames_to_column(., "sample") %>% 
  merge(., sample_table, by="sample")


```

#Genome-specific GIFT profiles

```{r genome_level_GIFTs}

GIFTs_elements %>%
  as_tibble(., rownames = "MAG") %>%
  reshape2::melt() %>%
  rename(Code_element = variable, GIFT = value) %>%
  inner_join(GIFT_db,by="Code_element") %>%
  ggplot(., aes(x=Code_element, y=MAG, fill=GIFT, group=Code_function))+
    geom_tile()+
    scale_y_discrete(guide = guide_axis(check.overlap = TRUE))+
    scale_x_discrete(guide = guide_axis(check.overlap = TRUE))+
    scale_fill_gradientn(limits = c(0,1), colours=brewer.pal(7, "YlGnBu"))+
    facet_grid(. ~ Code_function, scales = "free", space = "free")+
    theme_grey(base_size=8)+
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1),strip.text.x = element_text(angle = 90))

```



#Element-level community-averaged GIFT profiles 

```{r community_averaged_GIFTs}
GIFTs_elements_community %>%
  reshape2::melt() %>%
  rename(Sample = Var1, Code_element = Var2, GIFT = value) %>%
  left_join(GIFT_db,by="Code_element") %>%
  ggplot(., aes(x=Code_element, y=Sample, fill=GIFT))+
    geom_tile()+
    scale_y_discrete(guide = guide_axis(check.overlap = TRUE))+
    scale_x_discrete(guide = guide_axis(check.overlap = TRUE))+
    scale_fill_gradientn(colours=brewer.pal(7, "YlGnBu"))+
    theme_grey(base_size=8)+
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1),strip.text.x = element_text(angle = 90))
```

```{r nmds}
GIFTs_elements_nmds <- GIFTs_elements_community %>%
  dist() %>%
  metaMDS(.,trymax = 500, k=2, verbosity=FALSE) %>%
  vegan::scores() %>%
  as_tibble(., rownames = "sample") %>%
  left_join(sample_table, by = join_by(sample == sample))


GIFTs_elements_nmds %>%
  group_by(species) %>% 
  mutate(x_cen = mean(NMDS1, na.rm = TRUE)) %>%
  mutate(y_cen = mean(NMDS2, na.rm = TRUE)) %>%
  ungroup() %>%
  ggplot(., aes(x=NMDS1,y=NMDS2, color=species, shape=Area_type)) +
  scale_color_manual(values=squirrel_colors) +
  geom_point(size=2.5) +
  geom_segment(aes(x=x_cen, y=y_cen, xend=NMDS1, yend=NMDS2), alpha=0.2) +
  theme_classic() +
  theme(legend.position="right", legend.box="vertical") +
  guides(color=guide_legend(title="Species"), shape=guide_legend(title="Area type")) 
```



# Function-level community-averaged GIFT profiles

```{r function_community_GIFTs}
# sample_sort <- sample_table %>%
#   select(sample,species,Area_type) %>%
#   arrange(species,Area_type) %>%
#   pull()
   
GIFTs_functions_community %>%
  reshape2::melt() %>%
  rename(sample = Var1, Code_function = Var2, GIFT = value) %>%
  left_join(GIFT_db,by = join_by(Code_function == Code_function)) %>%
  left_join(sample_table, by = join_by(sample == sample)) %>%
  #mutate(sample=factor(Sample, levels = sample_sort)) %>%
  ggplot(., aes(x=Code_function, y=sample, fill=GIFT))+
    geom_tile()+
    scale_y_discrete(guide = guide_axis(check.overlap = TRUE))+
    scale_x_discrete(guide = guide_axis(check.overlap = TRUE))+
    scale_fill_gradientn(colours=brewer.pal(7, "YlGnBu"))+
    #facet_grid(~species+Area_type, scales="free", space="free") +
    theme_grey(base_size=8)+
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1),strip.text.x = element_text(angle = 90))

```



##Domain-level

```{r}
alpha_colors <- c("lightblue", "pink", "lightgreen", "orange","lightyellow" )
group_n <- length(unique(merge_gift$species))

squirrel_colors <- c("#999999", "#cc3333")

#Biosynthesis by species
merge_gift %>%
  ggboxplot(., x = "species", y = "Biosynthesis", color = "species", fill="white", add="jitter") +
      scale_color_manual(values=squirrel_colors) +
      scale_fill_manual(values=paste0(squirrel_colors)) +
      stat_compare_means() +
      theme_classic() +
      labs(y = "Biosynthesis functions") +
      theme(
        legend.position = "none",
        axis.title.x = element_blank())

#Degradation by species
merge_gift %>%
  ggboxplot(., x = "species", y = "Degradation", color = "species", fill="white", add="jitter") +
      scale_color_manual(values=squirrel_colors) +
      scale_fill_manual(values=paste0(squirrel_colors)) +
      stat_compare_means() +
      theme_classic() +
      labs(y = "Degradation functions") +
      theme(
        legend.position = "none",
        axis.title.x = element_blank())

#Structure by species
merge_gift %>%
  ggboxplot(., x = "species", y = "Structure", color = "species", fill="white", add="jitter") +
      scale_color_manual(values=squirrel_colors) +
      scale_fill_manual(values=paste0(squirrel_colors)) +
      stat_compare_means() +
      theme_classic() +
      labs(y = "Structural functions") +
      theme(
        legend.position = "none",
        axis.title.x = element_blank())

merge_gift$Area_type <-factor(merge_gift$Area_type, levels = c("natural", "seminat", "urban"))
area_colors <- c("#76b183","#d57d2c","#6b7398")

#Biosynthesis by species*area_type
merge_gift %>%
  ggboxplot(., x = "species", y = "Biosynthesis", color = "Area_type", fill="white", add="jitter") +
  scale_color_manual(values=area_colors) +
  scale_fill_manual(values=paste0(area_colors)) +
  stat_compare_means() +
  theme_classic() +
  labs(y = "Biosynthesis") +
  theme(
    legend.position = "top",
    legend.box = "horizontal",
    axis.title.x = element_blank()) +
  guides(color=guide_legend(title="Area type"), fill="none")

#Degradation by species*area_type
merge_gift %>%
  ggboxplot(., x = "species", y = "Degradation", color = "Area_type", fill="white", add="jitter") +
  scale_color_manual(values=area_colors) +
  scale_fill_manual(values=paste0(area_colors)) +
  stat_compare_means() +
  theme_classic() +
  labs(y = "Biosynthesis") +
  theme(
    legend.position = "top",
    legend.box = "horizontal",
    axis.title.x = element_blank()) +
  guides(color=guide_legend(title="Area type"), fill="none")

#Structure by species*area_type
merge_gift %>%
  ggboxplot(., x = "species", y = "Structure", color = "Area_type", fill="white", add="jitter") +
  scale_color_manual(values=area_colors) +
  scale_fill_manual(values=paste0(area_colors)) +
  stat_compare_means() +
  theme_classic() +
  labs(y = "Biosynthesis") +
  theme(
    legend.position = "top",
    legend.box = "horizontal",
    axis.title.x = element_blank()) +
  guides(color=guide_legend(title="Area type"), fill="none")



### Differences in bacterial functional capacity
#grid.arrange(arrangeGrob(p1, p5,p3, p4, ncol = 2))

```
