==> 06-DA analysis.Rmd <==
---
title: "DA analysis"
author: "Claudia Romeo"
date: "2023-12-04"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r install/load_libraries, echo=FALSE}

# install.packages(c("modeest", "lmerTest", "foreach", "parallel", "ggplot2", "ggrepel"))
# devtools::install_github("joey711/phyloseq")
# devtools::install_github("zhouhj1994/LinDA")

library(phyloseq)
library(LinDA)
library(GUniFrac) #ZicoSeq
suppressPackageStartupMessages(library(ALDEx2))
suppressPackageStartupMessages(library(ANCOMBC))

```

```{r prepare_phyloseq_objects}

otu_mat <- as.matrix(count_table_cov_size)
tax_mat <- as.matrix(mags_table)
samples_df <- as.data.frame(sample_table) %>% 
    tibble::column_to_rownames("sample") %>%
    select(species,sqID,Area_type)

# otu = otu_table(otu_mat, taxa_are_rows = TRUE)
# tax = tax_table(tax_mat)
# samples = sample_data(samples_df)
# 
# squirrels_ps <- phyloseq(otu, tax, samples)

```

```{r LinDA}

linda.obj <- linda(otu_mat, samples_df, formula = '~ species+(1|sqID)', alpha = 0.05,
                   prev.cut = 0.1, lib.cut = 1000, winsor.quan = 0.97)

linda.plot(linda.obj, c('speciesSciurusvulgaris'),
           titles = c('species'), alpha = 0.05, lfc.cut = 1,
           legend = TRUE, directory = NULL, width = 11, height = 8)

L <- matrix(c(0, 1), nrow = 1, byrow = TRUE)
linda.wald.test(linda.obj, L, 'LMM', alpha = 0.05)


# #install package "phyloseq" for importing "smokers" dataset
# ind <- smokers$meta$AIRWAYSITE == 'Throat'
# otu.tab <- as.data.frame(smokers$otu[, ind])

# meta <- cbind.data.frame(Smoke = factor(smokers$meta$SMOKER[ind]),
#                          Sex = factor(smokers$meta$SEX[ind]),
#                          Site = factor(smokers$meta$SIDEOFBODY[ind]),
#                          SubjectID = factor(smokers$meta$HOST_SUBJECT_ID[ind]))

# ind1 <- which(meta$Site == 'Left')
# res.left <- linda(otu.tab[, ind1], meta[ind1, ], formula = '~Smoke+Sex', alpha = 0.1,
#                   prev.cut = 0.1, lib.cut = 1000, winsor.quan = 0.97)

# ind2 <- which(meta$Site == 'Right')
# res.right <- linda(otu.tab[, ind2], meta[ind2, ], formula = '~Smoke+Sex', alpha = 0.1,
#                    prev.cut = 0.1, lib.cut = 1000, winsor.quan = 0.97)

# rownames(res.left$output[[1]])[which(res.left$output[[1]]$reject)]
# rownames(res.right$output[[1]])[which(res.right$output[[1]]$reject)]
# 
# linda.obj <- linda(otu.tab, meta, formula = '~Smoke+Sex+(1|SubjectID)', alpha = 0.1,
#                    prev.cut = 0.1, lib.cut = 1000, winsor.quan = 0.97)
# linda.plot(linda.obj, c('Smokey', 'Sexmale'), 
#            titles = c('Smoke: n v.s. y', 'Sex: female v.s. male'), alpha = 0.1, lfc.cut = 1,
#            legend = TRUE, directory = NULL, width = 11, height = 8)
# 
# L <- matrix(c(0, 1, 0, 0, 0, 1), nrow = 2, byrow = TRUE)
# #L <- matrix(c(0, 1, 0), nrow = 1, byrow = TRUE)
# linda.wald.test(linda.obj, L, 'LMM', alpha = 0.1)


```

```{r ZicoSeq}
# ZicoSeq.obj <- ZicoSeq(meta.dat = meta.dat, feature.dat = comm, 
#                     grp.name = 'SmokingStatus', adj.name = 'Sex', feature.dat.type = "count",
#                     # Filter to remove rare taxa
#                     prev.filter = 0.2, mean.abund.filter = 0,  
#                     max.abund.filter = 0.002, min.prop = 0, 
#                     # Winsorization to replace outliers
#                     is.winsor = TRUE, outlier.pct = 0.03, winsor.end = 'top',
#                     # Posterior sampling 
#                     is.post.sample = TRUE, post.sample.no = 25, 
#                     # Use the square-root transformation
#                     link.func = list(function (x) x^0.5), stats.combine.func = max,
#                     # Permutation-based multiple testing correction
#                     perm.no = 99,  strata = NULL, 
#                     # Reference-based multiple stage normalization
#                     ref.pct = 0.5, stage.no = 6, excl.pct = 0.2,
#                     # Family-wise error rate control
#                     is.fwer = TRUE, verbose = TRUE, return.feature.dat = TRUE)

count_table_rounded <- round(count_table_cov_size, digits=0)
otu_mat2 <- as.matrix(count_table_rounded)


ZicoSeq.obj <- ZicoSeq(meta.dat = samples_df, feature.dat = otu_mat2, 
                       grp.name = 'species', adj.name = 'Area_type', feature.dat.type ="count",
                       # Filter to remove rare taxa
                      prev.filter = 0.2, mean.abund.filter = 0,  
                      max.abund.filter = 0.002, min.prop = 0, 
                      # Winsorization to replace outliers
                      is.winsor = TRUE, outlier.pct = 0.03, winsor.end = 'top',
                      # Posterior sampling 
                      is.post.sample = TRUE, post.sample.no = 25, 
                      # Use the square-root transformation
                      link.func = list(function (x) x^0.5), stats.combine.func = max,
                      # Permutation-based multiple testing correction
                      perm.no = 99,  strata = NULL, 
                      # Reference-based multiple stage normalization
                      ref.pct = 0.5, stage.no = 6, excl.pct = 0.2,
                      # Family-wise error rate control
                      is.fwer = TRUE, verbose = TRUE, return.feature.dat = TRUE)

ZicoSeq.plot(ZicoSeq.obj, pvalue.type = 'p.adj.fdr', cutoff = 0.1, text.size = 10,
             out.dir = NULL, width = 10, height = 6)
```

```{r ALDEx2}

# if (!require("BiocManager", quietly = TRUE))
#     install.packages("BiocManager")
# 
# BiocManager::install("ALDEx2")
# BiocManager::install("ANCOMBC")




genome_counts <-rownames_to_column(genome_counts, var = "genome")

Sv_samples <- sample_table %>% 
                    #filter(type == "digesta") %>% #filter only faecal samples
                    filter(species == "Sciurus vulgaris") %>% 
                    dplyr::select(sample) %>% pull()

Sc_samples <- sample_table %>% 
                    #filter(type == "digesta") %>% #filter only faecal samples
                    filter(species == "Sciurus carolinensis") %>% 
                    dplyr::select(sample) %>% pull()

structural_zeros <- genome_counts %>% 
   rowwise() %>% #compute for each row (genome)
   mutate(all_zeros_Sv = all(c_across(all_of(Sv_samples)) == 0)) %>% # set true if all samples in Sv have zeros
   mutate(all_zeros_Sc = all(c_across(all_of(Sc_samples)) == 0)) %>% # set true if all samples in Sc have zeros
   mutate(average_Sv = mean(c_across(all_of(Sv_samples)), na.rm = TRUE)) %>% # get average genome counts across Sv
   mutate(average_Sc = mean(c_across(all_of(Sc_samples)), na.rm = TRUE)) %>% # get average genome counts across Sc
   filter(all_zeros_Sv == TRUE || all_zeros_Sc==TRUE)  %>% # filter only genomes with structural zeros
   mutate(present = case_when(
      all_zeros_Sv & !all_zeros_Sc ~ "Sc",
      !all_zeros_Sv & all_zeros_Sc ~ "Sv",
      !all_zeros_Sv & !all_zeros_Sc ~ "None",
      TRUE ~ NA_character_
    )) %>%
   mutate(average = ifelse(present == "Sv", average_Sv, average_Sc)) %>%
   dplyr::select(genome, present, average) %>%
   left_join(mags_table, by=join_by(genome==genome)) %>%
   arrange(present,-average)

structural_zeros %>%
    mutate(average = ifelse(present == "Sv", average * -1, average)) %>% #convert TJ1 genome counts to negative
    ggplot(., aes(x=average, y=forcats::fct_rev(phylum), color=phylum)) +
      geom_jitter(size=3) +
      geom_vline(xintercept=0) + 
      xlim(-max(structural_zeros$average)-3,max(structural_zeros$average)+3) +
      scale_color_manual(values=phylum_colors) +
      geom_text(aes(-max(structural_zeros$average)+5, 1), label = "Only present\nin Sv", color="#666666") +
      geom_text(aes(max(structural_zeros$average)-5, 1), label = "Only present\nin Sc", color="#666666") +
     theme(legend.position='none',
          panel.background = element_blank(),
          axis.line = element_line(size = 0.5, linetype = "solid", colour = "black"))+
      labs(y="Genus",x="Genome counts") + 
      guides(col=guide_legend("Phylum"))

```


```{r}
genome_counts_filt <- genome_counts %>%
  filter(!genome %in% structural_zeros$genome) %>% # remove structural zeros
  column_to_rownames(var="genome") %>%
  mutate_all(~ . * 1e6) %>% #multiple by a million
  round(0) #round to integer

genome_counts_filt.clr <- aldex.clr(genome_counts_filt, 
               sample_table$species, 
               mc.samples=128, 
               denom="all", 
               verbose=F)

genome_counts_filt.ttest <- aldex.ttest(genome_counts_filt.clr, 
                hist.plot=F, 
                paired.test=F, 
                verbose=F)

genome_counts_filt.effect <- aldex.effect(genome_counts_filt.clr, 
              CI=T, 
              verbose=F, 
              include.sample.summary=F, 
              glm.conds=NULL, 
              useMC=F)

genome_counts_filt.all <- data.frame(genome_counts_filt.ttest,genome_counts_filt.effect) %>%
    rownames_to_column(var="genome") %>%
    left_join(mags_table,by=join_by(genome==genome))

genome_counts_filt.all %>%
    mutate(significance = ifelse(wi.eBH < 0.05, "1", "0")) %>%
    ggplot(., aes(x=rab.all, y=diff.btw, color=significance)) +
      geom_point() +
      scale_color_manual(values = c("#cccccc","#00FFFF")) +
      labs(color="Significance", y="Difference between species", x="Mean CLR") +
      theme_classic()

genome_counts_filt.all %>%
    mutate(significance = ifelse(wi.eBH < 0.05, "1", "0")) %>%
    ggplot(., aes(x=-log(wi.eBH), y=diff.btw, color=significance)) +
      geom_point() +
      scale_color_manual(values = c("#cccccc","#00FFFF")) +
      labs(color="Significance", y="Difference between species", x="p-value") +
      theme_classic()

genome_counts_filt.all %>%
    mutate(significance = ifelse(wi.eBH < 0.05, "1", "0")) %>%
    ggplot(., aes(x=diff.win, y=diff.btw, color=significance)) +
      geom_abline(intercept = 0, slope =  0, size=0.4, linetype="solid",  color="#000000") +
      geom_abline(intercept = 0, slope =  1, size=0.4, linetype="dashed", color="#000000") +
      geom_abline(intercept = 0, slope =  2, size=0.6,   linetype="dashed", color="#000000") +
      geom_abline(intercept = 0, slope =  4, size=0.8,   linetype="dashed", color="#000000") +
      geom_abline(intercept = 0, slope = -1, size=0.4, linetype="dashed", color="#000000") + 
      geom_abline(intercept = 0, slope = -2, size=0.6,   linetype="dashed", color="#000000") +
      geom_abline(intercept = 0, slope = -4, size=0.8,   linetype="dashed", color="#000000") +
      geom_point() +
      scale_color_manual(values = c("#cccccc90","#00FFFF90")) +
      geom_text(aes(2.5, 20), label = "Enriched\nin Sc", color="#666666") +
      geom_text(aes(2.5, -20), label = "Enriched\nin Sv", color="#666666") +
      labs(color="Significance", y="Difference between species", x="Dispersion within species") +
      theme_classic()

genome_counts_filt.all %>%
    mutate(significance = ifelse(overlap < 0.1, "1", "0")) %>%
    ggplot(., aes(x=diff.win, y=diff.btw, color=significance)) +
      geom_abline(intercept = 0, slope =  0, size=0.4, linetype="solid",  color="#000000") +
      geom_abline(intercept = 0, slope =  1, size=0.4, linetype="dashed", color="#000000") +
      geom_abline(intercept = 0, slope =  2, size=0.6,   linetype="dashed", color="#000000") +
      geom_abline(intercept = 0, slope =  4, size=0.8,   linetype="dashed", color="#000000") +
      geom_abline(intercept = 0, slope = -1, size=0.4, linetype="dashed", color="#000000") + 
      geom_abline(intercept = 0, slope = -2, size=0.6,   linetype="dashed", color="#000000") +
      geom_abline(intercept = 0, slope = -4, size=0.8,   linetype="dashed", color="#000000") +
      geom_point() +
      scale_color_manual(values = c("#cccccc90","#00FFFF90")) +
      geom_text(aes(2.5, 20), label = "Enriched\nin Sc", color="#666666") +
      geom_text(aes(2.5, -20), label = "Enriched\nin Sv", color="#666666") +
      labs(color="Significance", y="Difference between species", x="Dispersion within species") +
      theme_classic()

#Get phylum colors from the EHI standard
phylum_colors <- ehi_phylum_colors %>%
  right_join(mags_table, by=join_by(phylum == phylum)) %>%
    arrange(match(genome, tree$tip.label)) %>%
    dplyr::select(phylum, colors) %>%
    unique() %>%
    arrange(phylum) %>%
    dplyr::select(colors) %>%
    pull()

genome_counts_filt.all %>%
    mutate(significance = ifelse(wi.eBH < 0.05, "1", "0")) %>%
    ggplot(., aes(x=diff.win, y=diff.btw, color=phylum, shape=significance)) +
      geom_abline(intercept = 0, slope =  0, size=0.4, linetype="solid",  color="#000000") +
      geom_abline(intercept = 0, slope =  1, size=0.4, linetype="dashed", color="#000000") +
      geom_abline(intercept = 0, slope =  2, size=0.6,   linetype="dashed", color="#000000") +
      geom_abline(intercept = 0, slope =  4, size=0.8,   linetype="dashed", color="#000000") +
      geom_abline(intercept = 0, slope = -1, size=0.4, linetype="dashed", color="#000000") + 
      geom_abline(intercept = 0, slope = -2, size=0.6,   linetype="dashed", color="#000000") +
      geom_abline(intercept = 0, slope = -4, size=0.8,   linetype="dashed", color="#000000") +
      geom_point() +
      scale_color_manual(values=phylum_colors) +
      scale_shape_manual(values=c(1,19))+
      geom_text(aes(2.5, 20), label = "Enriched\nin Sc", color="#666666") +
      geom_text(aes(2.5, -20), label = "Enriched\nin Sv", color="#666666") +
      labs(color="Phylum", shape="Significance", y="Difference between species", x="Dispersion within species") +
      theme_classic()
```

```{r ANCOMBC}


phylo_samples <- sample_table %>% 
                    column_to_rownames("sample") %>% 
                    mutate(species2 = factor(ifelse(species=="Sciurus carolinensis", "Sc", "Sv"))) %>%
                    sample_data() #convert to phyloseq sample_data object

phylo_counts <- genome_counts %>% 
                    filter(!genome %in% structural_zeros$genome) %>% # remove structural zeros
                    column_to_rownames("genome") %>% 
                    mutate_all(~ replace(., . == 0, 0.00001)) %>% #add pseudo counts to avoid structural zero issues (note this approach can be improved!)
                    otu_table(., taxa_are_rows = TRUE) #convert to phyloseq otu_table object

phylo_taxonomy <- mags_table %>% 
                    filter(genome %in% rownames(phylo_counts)) %>% # remove structural zeros
                    rownames_to_column("genome1") %>%
                    mutate(genome2=genome) %>% #create a pseudo genome name column
                    column_to_rownames("genome2") %>% 
                    dplyr::select(domain,phylum,class,order,family,genus,species,genome) %>% #add an additional taxonomic level to ensure genome-level analysis (as no all genomes have species-level taxonomic assignments. Otherwise, ANCOMBC2 aggregates analyses per species)
                    as.matrix() %>% 
                    tax_table() #convert to phyloseq tax_table object

#Generate phyloseq object required to input ANCOMBC
genome_data <- phyloseq(phylo_counts, phylo_taxonomy, phylo_samples)
```


```{r}
set.seed(1234) #set seed for reproducibility
ancom_output = ancombc2(data = genome_data, 
                  assay_name = "counts", 
                  tax_level = NULL, #change to agglomerate analysis to a higher taxonomic range
                  fix_formula = "species2", #fixed variable(s)
                  p_adj_method = "holm", 
                  pseudo_sens = TRUE,
                  prv_cut = 0.10, 
                  lib_cut = 1000, 
                  s0_perc = 0.05,
                  group = NULL, 
                  struc_zero = FALSE, 
                  neg_lb = FALSE,
                  alpha = 0.05, 
                  n_cl = 2, 
                  verbose = TRUE,
                  global = FALSE, 
                  pairwise = FALSE, 
                  dunnet = FALSE, 
                  trend = FALSE,
                  iter_control = list(tol = 1e-5, max_iter = 20, verbose = FALSE),
                  em_control = list(tol = 1e-5, max_iter = 100),
                  lme_control = NULL, 
                  mdfdr_control = list(fwer_ctrl_method = "holm", B = 100), 
                  trend_control = NULL)

ancom_result <- ancom_output$res %>%
    rename(genome=taxon) %>%
    left_join(mags_table,by=join_by(genome==genome))
ancom_result %>%
    mutate(significance = ifelse(p_species2Sv < 0.01, "1", "0")) %>%
    ggplot(., aes(x=-log(p_species2Sv), y=lfc_species2Sv, color=significance)) +
      geom_point() +
      scale_color_manual(values = c("#cccccc","#00FFFF")) +
      geom_text(aes(2.5, 10), label = "Enriched\nin Sv", color="#666666") +
      geom_text(aes(2.5, -10), label = "Enriched\nin Sc", color="#666666") +
      labs(color="Significance", y="Difference between species", x="p-value") +
      theme_classic()
```
```{r}
phylo_counts %>% 
    as.data.frame() %>%
    mutate(average = rowMeans(across(everything())))  %>%
    rownames_to_column(var="genome") %>% 
    right_join(ancom_result, by=join_by(genome==genome)) %>% 
    mutate(significance = ifelse(p_species2Sv < 0.01, "1", "0")) %>%
    ggplot(., aes(x=average, y=lfc_species2Sv, color=significance)) +
      geom_point() +
      scale_color_manual(values = c("#cccccc","#00FFFF")) +
      geom_text(aes(150, 10), label = "Enriched\nin Sv", color="#666666") +
      geom_text(aes(150, -10), label = "Enriched\nin Sc", color="#666666") +
      labs(color="Significance", y="Difference between treatments", x="Average genome count") +
      theme_classic()
```
```{r}
#Get phylum colors from the EHI standard
phylum_colors <- ehi_phylum_colors %>%
  right_join(ancom_result, by=join_by(phylum == phylum)) %>%
    arrange(match(genome, tree$tip.label)) %>%
    dplyr::select(phylum, colors) %>%
    mutate(colors = str_c(colors, "80"))  %>% #add 80% alpha
    unique() %>%
    arrange(phylum) %>%
    dplyr::select(colors) %>%
    pull()

ancom_result %>%
    filter(p_species2Sv < 0.01) %>%
    ggplot(., aes(x=lfc_species2Sv, y=forcats::fct_rev(phylum), color=phylum)) +
      geom_jitter(size=3) +
      geom_vline(xintercept=0) + 
      xlim(-max(ancom_result$lfc_species2Sv)-2,max(ancom_result$lfc_species2Sv)+2) +
      scale_color_manual(values=phylum_colors) +
      geom_text(aes(-max(ancom_result$lfc_tspecies2Sv)+1, 1), label = "Enriched\nin Sv", color="#666666") +
      geom_text(aes(max(ancom_result$lfc_species2Sv)-1, 1), label = "Enriched\nin Sc", color="#666666") +
     theme(legend.position='none',
          panel.background = element_blank(),
          axis.line = element_line(size = 0.5, linetype = "solid", colour = "black"),
          axis.title.x=element_blank())+
      labs(y="Genus",x="Genome counts") + 
      guides(col=guide_legend("Phylum"))
```
