==> 08-hmsc_setup_new.Rmd <==


# HMSCs - model setup

```{r load_data08}
rm(list=ls()) #clear environment
load("data/squirrels_data.Rdata")
singlem <- read.csv("data/singlem.csv",sep=";",header=T)

red_samples <- sample_metadata %>%
  right_join(singlem, by = join_by(sample == sample)) %>%
  filter(species == "Sciurus vulgaris",
         development=="Adult",  #remove juveniles, nursing and pregnant females
         est_mapp>0.8) %>%  #remove samples with low DAMR
  dplyr::select(sample) %>%
  pull()

grey_samples <- sample_metadata %>%
  right_join(singlem, by = join_by(sample == sample)) %>%
  filter(species == "Sciurus carolinensis",
         development=="Adult",  #remove juveniles, nursing and pregnant females
         est_mapp>0.8) %>%  #remove samples with low DAMR
  dplyr::select(sample) %>%
  pull()

```

## Data preparation

```{r subsetting02, warning=FALSE, comments="", message=FALSE}


# Subset by prevalence (present in more than 5 samples)
selected_genomes_red0 <- genome_counts %>%
    column_to_rownames(var="genome") %>%
    select(any_of(red_samples)) %>%
    filter(rowSums(across(starts_with("EHI")) != 0) >= 5) %>%
    rownames()

selected_genomes_grey0 <- genome_counts %>%
    column_to_rownames(var="genome") %>%
    select(any_of(grey_samples)) %>%
    filter(rowSums(across(starts_with("EHI")) != 0) >= 5) %>%
    rownames()

# Subset by minimum representation of 1% relative abundance in at least 5 samples
selected_genomes_red <- genome_counts %>%
    filter(genome %in% selected_genomes_red0) %>%
    column_to_rownames(var="genome") %>%
    hilldiv2::tss() %>%
    as.data.frame() %>%
    filter(rowSums(across(starts_with("EHI")) >= 0.01) >= 5) %>%
    rownames()

selected_genomes_grey <- genome_counts %>%
    filter(genome %in% selected_genomes_grey0) %>%
    column_to_rownames(var="genome") %>%
    hilldiv2::tss() %>%
    as.data.frame() %>%
    filter(rowSums(across(starts_with("EHI")) >= 0.01) >= 5) %>%
    rownames()

# # Subset genome metadata
# genome_metadata_red <- genome_metadata %>%
#     filter(genome %in% selected_genomes_red)
# 
# genome_metadata_grey <- genome_metadata %>%
#     filter(genome %in% selected_genomes_grey)

```

## Prepare input for HMSCs

### Red squirrel

```{r hmsc_input_red, warning=FALSE, comments="", message=FALSE}

#RED SQUIRREL MODEL

# Random effects data (study design)
StudyDesign.red <- sample_metadata %>%
                    filter(species=="Sciurus vulgaris") %>%
                    select(sample,animal,sampling_site) %>%
                    mutate(sampling_site = factor(sampling_site)) %>%
                    mutate(animal = factor(animal)) %>%
                    column_to_rownames("sample")

#Calculate normalisation factor to account for genome length
normalisation_factor <- genome_metadata %>% 
  filter(genome %in% selected_genomes_red) %>%
  mutate(factor=median(length)/length) %>%
  pull(factor)

# Genome count table (quantitative community data)
YData.red <- read_counts  %>%
                    filter(genome %in% selected_genomes_red) %>% #subset genomes
                    mutate(across(where(is.numeric), ~ round(. * normalisation_factor,0) )) %>% 
                    mutate(across(where(is.numeric), ~ . +1 )) %>% #add +1 pseudocount to remove zeros
                    mutate(across(where(is.numeric), ~  log(.) )) %>% #log-transform
                    column_to_rownames("genome") %>%
                    select(all_of(row.names(StudyDesign.red))) %>%  
                    as.data.frame() %>%
                    t() # transpose

# Fixed effects data (explanatory variables) 
XData.red <- sample_metadata %>%
                    filter(species=="Sciurus vulgaris") %>%
                    select(sample,sex,index500,season) %>%
                    mutate(logseqdepth=read_counts %>% #total log-sequencing depth
                        select(all_of(row.names(StudyDesign.red))) %>%
                        colSums() %>%
                        log()
                    ) %>%
                    mutate(index500 = index500) %>%
                    mutate(season = factor(season, levels=c("spring-summer","autumn","winter"))) %>%
                    mutate(sex = factor(sex)) %>%
                    column_to_rownames("sample")


# Genome trait data
TrData.red <- genome_gifts %>% rownames_to_column(var="genome") %>%
                    filter(genome %in% selected_genomes_red) %>% #subset genomes
                    arrange(match(genome, colnames(YData.red))) %>%
                    column_to_rownames(var="genome") %>%
                    to.elements(.,GIFT_db) %>%
                    to.functions(.,GIFT_db) %>%
                    as.data.frame()

# Genome phylogeny
PData <- genome_tree
```

### Grey squirrel

```{r hmsc_input_grey, warning=FALSE, comments="", message=FALSE}

#GREY SQUIRREL MODEL

# Random effects data (study design)
StudyDesign.grey <- sample_metadata %>%
                    filter(species=="Sciurus carolinensis") %>%
                    select(sample,animal,sampling_site) %>%
                    mutate(sampling_site = factor(sampling_site)) %>%
                    mutate(animal = factor(animal)) %>%
                    column_to_rownames("sample")

#Calculate normalisation factor to account for genome length
normalisation_factor <- genome_metadata %>% 
  filter(genome %in% selected_genomes_grey) %>%
  mutate(factor=median(length)/length) %>%
  pull(factor)

# Genome count table (quantitative community data)
YData.grey <- read_counts  %>%
                    filter(genome %in% selected_genomes_grey) %>% #subset genomes
                    mutate(across(where(is.numeric), ~ round(. * normalisation_factor,0) )) %>% 
                    mutate(across(where(is.numeric), ~ . +1 )) %>% #add +1 pseudocount to remove zeros
                    mutate(across(where(is.numeric), ~  log(.) )) %>% #log-transform
                    column_to_rownames("genome") %>%
                    select(all_of(row.names(StudyDesign.grey))) %>%  
                    as.data.frame() %>%
                    t() # transpose

# Fixed effects data (explanatory variables) 
XData.grey <- sample_metadata %>%
                    filter(species=="Sciurus carolinensis") %>%
                    select(sample,sex,index500,season) %>%
                    mutate(logseqdepth=read_counts %>% #total log-sequencing depth
                        select(all_of(row.names(StudyDesign.grey))) %>%
                        colSums() %>%
                        log()
                    ) %>%
                    mutate(index500 = index500) %>%
                    mutate(season = factor(season, levels=c("spring-summer","autumn","winter"))) %>%
                    mutate(sex = factor(sex)) %>%
                    column_to_rownames("sample")


# Genome trait data
TrData.grey <- genome_gifts %>% rownames_to_column(var="genome") %>%
                    filter(genome %in% selected_genomes_grey) %>% #subset genomes
                    arrange(match(genome, colnames(YData.grey))) %>%
                    column_to_rownames(var="genome") %>%
                    to.elements(.,GIFT_db) %>%
                    to.functions(.,GIFT_db) %>%
                    as.data.frame()

# Genome phylogeny
PData <- genome_tree
```

## Define HMSC models and MCMC

```{r hmsc_formulas, warning=FALSE, comments="", message=FALSE}

# Fixed effects formula
XFormula = ~sex + index500 + season + logseqdepth

# Study design
rL.animal.red = HmscRandomLevel(units = levels(StudyDesign.red$animal))
rL.sampling_site.red = HmscRandomLevel(units = levels(StudyDesign.red$sampling_site))

rL.animal.grey = HmscRandomLevel(units = levels(StudyDesign.grey$animal))
rL.sampling_site.grey = HmscRandomLevel(units = levels(StudyDesign.grey$sampling_site))

```


```{r hmsc_models, warning=FALSE, comments="", message=FALSE}
#Define models
model.red = Hmsc(Y=YData.red,
         XData = XData.red,
         XFormula = XFormula,
         studyDesign = StudyDesign.red,
         phyloTree = PData,
         ranLevels = list("animal"=rL.animal.red, "sampling_site"=rL.sampling_site.red),
         distr = "normal",
         YScale = TRUE)

model.grey = Hmsc(Y=YData.grey,
         XData = XData.grey,
         XFormula = XFormula,
         studyDesign = StudyDesign.grey,
         phyloTree = PData,
         ranLevels = list("animal"=rL.animal.grey, "sampling_site"=rL.sampling_site.grey),
         distr = "normal",
         YScale = TRUE)

#Save list of models as an R object.
model_list = list(model.red=model.red,model.grey=model.grey)
if (!dir.exists("hmsc")){dir.create("hmsc")}
save(model_list, file = "hmsc/hmsc.Rdata")

```

Upload **hmsc/hmsc.Rdata** to the HPC respecting the directory structure.



```{r hmsc_mcmc, warning=FALSE, comments="", message=FALSE}
# How often to sample the MCMC
MCMC_samples_list = 250

# The number of MCMC steps between each recording sample
MCMC_thin_list = c(1, 10)

# The number of MCMC chains to use
nChains = 4
```


## Generate HMSC executables

The next chunk generates shell files for every combination of model, MCMC samples and MCMM thinning, ready to be launched as SLURM jobs.

```{r hmsc_executables, warning=FALSE, comments="", message=FALSE, eval=FALSE}

modelchains <- expand.grid(model = names(model_list), sample = MCMC_samples_list, thin = MCMC_thin_list)

if (!dir.exists("hmsc")){dir.create("hmsc")}
for(i in c(1:nrow(modelchains))){
      modelname=as.character(modelchains[i,1])
      sample=modelchains[i,2]
      thin=modelchains[i,3]
      executablename <- paste0("hmsc/exe_",modelname,"_",sample,"_",thin,".sh")
      fitname <- paste0("hmsc/fit_",modelname,"_",sample,"_",thin,".Rdata")
      convname <- paste0("hmsc/conv_",modelname,"_",sample,"_",thin,".Rdata")
      model <- paste0('model_list$',modelname)
      psrf.beta.name <-  paste0("psrf.beta.",modelname,"_",sample,"_",thin)
      psrf.gamma.name <-  paste0("psrf.gamma.",modelname,"_",sample,"_",thin)
      psrf.rho.name <-  paste0("psrf.rho.",modelname,"_",sample,"_",thin)
      jobname <- paste0("hmsc_",modelname,"_",sample,"_",thin)
      minutes <- round(sample * thin * (ncol(YData.red)/50), 0)
      code <- sprintf("#!/bin/bash
#SBATCH --job-name=%s                   # Job name
#SBATCH --nodes=1
#SBATCH --ntasks=4                      # Run on 4 CPUs
#SBATCH --mail-user=antton.alberdi@sund.ku.dk
#SBATCH --mem=200gb                     # Job memory request
#SBATCH --time=%d                       # In minutes

# Activate conda environment
module load mamba/1.3.1
if ! conda info --envs | grep -q hmsc; then
  mamba create -p ./hmsc/hmsc_env -y r-essentials r-base r-tidyverse r-Hmsc
fi

# Activate conda environment
source activate ./hmsc/hmsc_env

# Run R script
Rscript -e '
library(tidyverse)
library(Hmsc)
# Load formulas and data
load(\"hmsc/hmsc.Rdata\")

# Declare placeholders
modelname = \"%s\"
model = %s
fitname = \"%s\"
convname = \"%s\"
sample = %d
thin = %d
nchains = %d

# Run model fitting
m = sampleMcmc(hM = model,
         samples = sample,
         thin = thin,
         adaptNf=rep(ceiling(0.4*sample*thin),model$nr),
         transient = ceiling(0.5*sample*thin),
         nChains = nchains,
         nParallel = nchains)

# Run model cross-validation
partition <- createPartition(m, nfolds = 5)
cv <- computePredictedValues(m, partition=partition, nChains = 4)

# Assess chain convergence
mpost = convertToCodaObject(m,
      spNamesNumbers = c(T,F),
      covNamesNumbers = c(T,F),
      Beta = TRUE,
      Gamma = TRUE,
      V = FALSE,
      Sigma = FALSE,
      Rho = TRUE,
      Eta = FALSE,
      Lambda = FALSE,
      Alpha = FALSE,
      Omega = FALSE,
      Psi = FALSE,
      Delta = FALSE) # Convert to CODA object

# Fixed effects
assign(paste0(\"psrf.beta.\", modelname,\"_\",sample,\"_\",thin), gelman.diag(mpost$Beta,multivariate=FALSE)$psrf)

# Traits
assign(paste0(\"psrf.gamma.\", modelname,\"_\",sample,\"_\",thin), gelman.diag(mpost$Gamma,multivariate=FALSE)$psrf)

# Phylogeny
assign(paste0(\"psrf.rho.\", modelname,\"_\",sample,\"_\",thin), gelman.diag(mpost$Rho,multivariate=FALSE)$psrf)

# Write convergence data
save(%s, %s, %s, file=convname)

# Save model fit object
save(m, cv, file=fitname)
'
", jobname, minutes, modelname, model, fitname, convname, sample, thin, nChains, psrf.beta.name, psrf.gamma.name, psrf.rho.name)
      writeLines(code, executablename)
    }
```

Upload the produced **hmsc/exe_XXXXX.sh** files to the HPC respecting the directory structure.

## Fit Hmsc models (in Mjolnir HPC)

Launch the SLURM jobs by using:

```{sh, eval=FALSE}
# Submit all .sh files in the hmsc folder
for jobfile in exe_*.sh; do
    sbatch "$jobfile"
done

#Or launch them one by one only the ones you want to launch
sbatch hmsc/exe_model.red_250_10.sh
sbatch hmsc/exe_model.red_250_10.sh

```

## Assess chain convergence

Convergence diagnostic values substantially above 1 indicate lack of convergence.
Values below 1.1 are considered good enough

```{r hmsc_convergence, warning=FALSE, comments="", message=FALSE, eval=TRUE}

# Load all conv file available in the hmsc folder
# list.files(path = "hmsc", pattern = "^conv", full.names = TRUE, include.dirs = TRUE) %>%
#   lapply(.,load,.GlobalEnv)

list.files(path = "hmsc", pattern = "^conv", full.names = TRUE, include.dirs = TRUE) %>%
  walk(~ {
    print(paste("Loading file:", .x))  # Debugging output: Print the file being loaded
    load(.x, envir = .GlobalEnv)
  })



# Create a merged psrf.beta (genome) plot
ls() %>%
        grep("^psrf\\.beta", ., value = TRUE) %>%
        map_dfr(~ {
         mat <- get(.x)
          data.frame(modelchain = .x, as.data.frame(mat, , stringsAsFactors = FALSE)) %>%
              rownames_to_column(var="parameter") %>%
              mutate(model = str_split(modelchain, "_") %>% map_chr(1) %>% gsub("psrf.beta.","",.)) %>%
              mutate(sample = str_split(modelchain, "_") %>% map_chr(2)) %>% #extract sample info from model name
              mutate(thin = str_split(modelchain, "_") %>% map_chr(3)) #extract thin info from model name
      }) %>%
      ggplot(.,aes(x=reorder(modelchain,-Point.est.,fun=function(x) {quantile(x, probs = 0.9)}),y=Point.est.)) +
        geom_violin(fill="#b8d9e3", color="#328da8") +
        geom_jitter(alpha=0.3,size=0.2, color="#a8babf") +
        stat_summary(fun=function(x) {quantile(x, probs = 0.9)}, geom="crossbar", width=0.2, color="orange") +
        geom_hline(yintercept=1.1, linetype="dashed", color = "red") +
        ylim(0.9,2)+
        labs(x="Model chains",y="Parameter estimates")+
        theme_classic()+
        coord_flip()

# Create a merged psrf.gamma (trait) plot
# ls() %>%
#         grep("^psrf\\.gamma", ., value = TRUE) %>%
#         map_dfr(~ {
#          mat <- get(.x)
#           data.frame(modelchain = .x, as.data.frame(mat, , stringsAsFactors = FALSE)) %>%
#               rownames_to_column(var="parameter") %>%
#               mutate(model = str_split(modelchain, "_") %>% map_chr(1) %>% gsub("psrf.gamma.","",.)) %>%
#               mutate(sample = str_split(modelchain, "_") %>% map_chr(2)) %>% #extract sample info from model name
#               mutate(thin = str_split(modelchain, "_") %>% map_chr(3)) #extract thin info from model name
#       }) %>%
#       ggplot(.,aes(x=reorder(modelchain,-Point.est.,fun=function(x) {quantile(x, probs = 0.9)}),y=Point.est.)) +
#         geom_violin(fill="#b8d9e3", color="#328da8") +
#         geom_jitter(alpha=0.3,size=0.2, color="#a8babf") +
#         stat_summary(fun=function(x) {quantile(x, probs = 0.9)}, geom="crossbar", width=0.2, color="orange") +
#         geom_hline(yintercept=1.1, linetype="dashed", color = "red") +
#         ylim(0.9,2)+
#         labs(x="Model chains",y="Parameter estimates")+
#         theme_classic()+
#         coord_flip()


# Create a merged psrf.rho (phylogeny) plot
# ls() %>%
#         grep("^psrf\\.rho", ., value = TRUE) %>%
#         map_dfr(~ {
#          mat <- get(.x)
#           data.frame(modelchain = .x, as.data.frame(mat, , stringsAsFactors = FALSE)) %>%
#               rownames_to_column(var="parameter") %>%
#               mutate(model = str_split(modelchain, "_") %>% map_chr(1) %>% gsub("psrf.rho.","",.)) %>%
#               mutate(sample = str_split(modelchain, "_") %>% map_chr(2)) %>% #extract sample info from model name
#               mutate(thin = str_split(modelchain, "_") %>% map_chr(3)) #extract thin info from model name
#       }) %>%
#       ggplot(.,aes(x=reorder(modelchain,-Point.est.,fun=function(x) {quantile(x, probs = 0.9)}),y=Point.est.)) +
#         geom_violin(fill="#b8d9e3", color="#328da8") +
#         geom_jitter(alpha=0.3,size=0.2, color="#a8babf") +
#         stat_summary(fun=function(x) {quantile(x, probs = 0.9)}, geom="crossbar", width=0.2, color="orange") +
#         geom_hline(yintercept=1.1, linetype="dashed", color = "red") +
#         ylim(0.9,2)+
#         labs(x="Model chains",y="Parameter estimates")+
#         theme_classic()+
#         coord_flip()

```

```{r model_fit, eval=FALSE}

# Select modelchain of interest
# load("hmsc/fit_model.red_250_10.Rdata")
# mred.WAIC <- computeWAIC(m, ghN = 10, byColumn = TRUE) %>% sum()
# 
# load("hmsc/fit_model.grey_250_10.Rdata")
# mgrey.WAIC <- computeWAIC(m, ghN = 10, byColumn = TRUE) %>% sum()
# 
# 
# WAICsum <- c(mred.WAIC,mgrey.WAIC) %>% rbind()
# 
# print(WAICsum)

# print(WAICmedian)
```
