==> 09-hmsc_analysis_new.Rmd <==


# HMSCs - output analysis

```{r load_data09, warning=FALSE, message=FALSE}

rm(list=ls()) #clear environment
load("data/squirrels_data.Rdata")
options(contrasts = c('contr.sum','contr.poly'))

# Select desired support threshold
support_threshold=0.9
negsupport_threshold=1-support_threshold

red_samples <- sample_metadata %>%
  filter(species == "Sciurus vulgaris") %>%
  dplyr::select(sample) %>%
  pull()

grey_samples <- sample_metadata %>%
  filter(species=="Sciurus carolinensis") %>%
  dplyr::select(sample) %>% 
  pull()

```

```{r hmsc_model_list, warning=FALSE, message=FALSE}

# Load modelchains 
load("hmsc/fit_model.red_250_10.Rdata")
m.red <- m
cv.red <- cv

load("hmsc/fit_model.grey_250_10.Rdata")
m.grey <- m
cv.grey <- cv

rm(m,cv)

levels <- c("sex", "index500", "season", "logseqdepth", "Random: animal", "Random: sampling_site")

# Basal tree
hmsc_tree_red <- genome_tree %>%
        keep.tip(., tip=m.red$spNames) 

hmsc_tree_grey <- genome_tree %>%
        keep.tip(., tip=m.grey$spNames) 

tree_all <- c(m.red$spNames,m.grey$spNames)

hmsc_tree_all <- genome_tree %>%
        keep.tip(., tip=tree_all) 

```

## Model fit

### Explanatory power

```{r model_expl_power, warning=FALSE, message=FALSE}

# Red squirrel
preds <- computePredictedValues(m.red) 
MF <- evaluateModelFit(hM=m.red, predY=preds) 
hist(MF$R2, xlim = c(0,1), main=paste0("Mean = ", round(mean(MF$R2),2)))

# Grey squirrel
preds <- computePredictedValues(m.grey) 
MF <- evaluateModelFit(hM=m.grey, predY=preds) 
hist(MF$R2, xlim = c(0,1), main=paste0("Mean = ", round(mean(MF$R2),2)))

```

### Predictive power

```{r hmsc_varpart, warning=FALSE, comments="", message=FALSE}

# Compute variance partitioning
varpart.red=computeVariancePartitioning(m.red)

varpart.grey=computeVariancePartitioning(m.grey)

```

```{r model_pred_power}

# Red squirrel
MFCV.red <- evaluateModelFit(hM=m.red, predY=cv.red)
mean(MFCV.red$R2, na.rm = TRUE)

var_pred_table.red <- tibble(mag=m.red$spNames,
       pred=MFCV.red$R2,
       var_pred=MFCV.red$R2 * varpart.red$vals[1,],
       support=getPostEstimate(hM=m.red, parName="Beta")$support %>% .[2,],
       estimate=getPostEstimate(hM=m.red, parName="Beta")$mean %>% .[2,]) 

genome_fit <- tibble(genome=m.red$spNames, r2 = MFCV.red[[2]])

# genome_counts %>%
#   select(genome, any_of(red_samples)) %>%
#   mutate_if(is.numeric, ~ . / sum(.)) %>%
#   left_join(genome_fit, by="genome") %>%
#   filter(r2>0.1) %>%
#   select(-c(genome,r2)) %>%
#   colSums() %>%
#   hist(main="R2", xlab="Proportion of microbiota")

genome_counts %>%
  select(genome, any_of(red_samples)) %>%
  mutate_if(is.numeric, ~ . / sum(.)) %>%
  left_join(var_pred_table.red, by=join_by("genome"=="mag")) %>%
  filter(var_pred>=0.005) %>%
  select(-c(genome,pred, var_pred, support, estimate)) %>%
  colSums() %>%
  hist(main="Predictive MAGs", xlab="Proportion of microbiota")


# Grey squirrel
MFCV.grey <- evaluateModelFit(hM=m.grey, predY=cv.grey)
mean(MFCV.grey$R2, na.rm = TRUE)

var_pred_table.grey <- tibble(mag=m.grey$spNames,
       pred=MFCV.grey$R2,
       var_pred=MFCV.grey$R2 * varpart.grey$vals[1,],
       support=getPostEstimate(hM=m.grey, parName="Beta")$support %>% .[2,],
       estimate=getPostEstimate(hM=m.grey, parName="Beta")$mean %>% .[2,]) 

genome_fit <- tibble(genome=m.grey$spNames, r2 = MFCV.grey[[2]])

genome_counts %>%
  select(genome, any_of(grey_samples)) %>%
  mutate_if(is.numeric, ~ . / sum(.)) %>%
  left_join(var_pred_table.grey, by=join_by("genome"=="mag")) %>%
  filter(var_pred>=0.005) %>%
  select(-c(genome,pred, var_pred, support, estimate)) %>%
  colSums() %>%
  hist(main="Predictive MAGs", xlab="Proportion of microbiota")




```


## Variance partitioning 

### Red squirrel


```{r hmsc_varpart_plot_red, warning=FALSE, comments="", message=FALSE, fig.dim=c(8, 10)}

# Varpart table
varpart.red$vals %>%
   as.data.frame() %>%
   rownames_to_column(var="variable") %>%
   pivot_longer(!variable, names_to = "genome", values_to = "value") %>%
   mutate(variable=factor(variable, levels=levels)) %>%
   group_by(variable) %>%
   summarise(mean=mean(value)*100,sd=sd(value)*100) %>%
   tt()

varpart_table.red <- varpart.red$vals %>%
   as.data.frame() %>%
   rownames_to_column(var="variable") %>%
   pivot_longer(!variable, names_to = "genome", values_to = "value") %>%
   mutate(variable=factor(variable, levels=rev(levels))) %>%
   mutate(genome=factor(genome, levels=rev(hmsc_tree_all$tip.label)))

# Phylums
phylums <- genome_metadata %>%
    filter(genome %in% hmsc_tree_all$tip.label) %>%
    arrange(match(genome, hmsc_tree_all$tip.label)) %>%
    mutate(phylum = factor(phylum, levels = unique(phylum))) %>%
    column_to_rownames(var = "genome") %>%
    select(phylum)

# Basal ggtree
varpart_tree <- hmsc_tree_all%>%
        force.ultrametric(.,method="extend") %>%
        ggtree(., size = 0.3)


# Add phylum colors next to the tree tips
varpart_tree <- gheatmap(varpart_tree, phylums, offset=-0.2, width=0.1, colnames=FALSE) +
   scale_fill_manual(values=custom_colors)+
      labs(fill="Phylum")

#Reset fill scale to use a different colour profile in the heatmap
varpart_tree <- varpart_tree + new_scale_fill()

# Add variance stacked barplot
vertical_tree_red <-  varpart_tree +
        scale_fill_manual(values=c("#122f3d","#83bb90","#cccccc","#ed8a45","#f6de9c","#b2b530","#be3e2b","#12738f")) +
        geom_fruit(
             data=varpart_table.red,
             geom=geom_bar,
             mapping = aes(x=value, y=genome, fill=variable, group=variable),
             pwidth = 2,
             offset = 0.05,
             width= 1,
             orientation="y",
             stat="identity")+
      labs(fill="Variable")


```


### Grey squirrel


```{r hmsc_varpart_plot_grey, warning=FALSE, comments="", message=FALSE, fig.dim=c(8, 10)}

# Varpart table
varpart.grey$vals %>%
   as.data.frame() %>%
   rownames_to_column(var="variable") %>%
   pivot_longer(!variable, names_to = "genome", values_to = "value") %>%
   mutate(variable=factor(variable, levels=levels)) %>%
   group_by(variable) %>%
   summarise(mean=mean(value)*100,sd=sd(value)*100) %>%
   tt()

varpart_table.grey <- varpart.grey$vals %>%
   as.data.frame() %>%
   rownames_to_column(var="variable") %>%
   pivot_longer(!variable, names_to = "genome", values_to = "value") %>%
   mutate(variable=factor(variable, levels=rev(levels))) %>%
   mutate(genome=factor(genome, levels=rev(hmsc_tree_grey$tip.label)))

# # Phylums
# phylums <- genome_metadata %>%
#     filter(genome %in% hmsc_tree_grey$tip.label) %>%
#     arrange(match(genome, hmsc_tree_grey$tip.label)) %>%
#     mutate(phylum = factor(phylum, levels = unique(phylum))) %>%
#     column_to_rownames(var = "genome") %>%
#     select(phylum)
# 
# # Basal ggtree
# varpart_tree <- hmsc_tree_grey %>%
#         force.ultrametric(.,method="extend") %>%
#         ggtree(., size = 0.3)
# 
# 
# # Add phylum colors next to the tree tips
# varpart_tree <- gheatmap(varpart_tree, phylums, offset=-0.2, width=0.1, colnames=FALSE) +
#    scale_fill_manual(values=custom_colors)+
#       labs(fill="Phylum")

#Reset fill scale to use a different colour profile in the heatmap
varpart_tree <- varpart_tree + new_scale_fill()

# Add variance stacked barplot
vertical_tree_both <-  vertical_tree_red +
        scale_fill_manual(values=c("#122f3d","#83bb90","#cccccc","#ed8a45","#f6de9c","#b2b530","#be3e2b","#12738f")) +
        geom_fruit(
             data=varpart_table.grey,
             geom=geom_bar,
             mapping = aes(x=value, y=genome, fill=variable, group=variable),
             pwidth = 2,
             offset = 0.05,
             width= 1,
             orientation="y",
             stat="identity")+
      labs(fill="Variable")

vertical_tree_both +
        vexpand(.07, -1) + # expand bottom
        annotate('text', x=4.8, y=-8, label = "S. vulgaris", color='black', fontface=2) +
        annotate('text', x=9.8, y=-8, label = "S. carolinensis", color='black', fontface=2)  

```


## Posterior estimates


```{r hmsc_postest, warning=FALSE, comments="", message=FALSE, fig.dim=c(8, 10)}

# Posterior estimates red
post_estimates_red <- getPostEstimate(hM=m.red, parName="Beta")$support %>%
    as.data.frame() %>%
    mutate(variable=m.red$covNames) %>%
    pivot_longer(!variable, names_to = "genome", values_to = "value") %>%
    mutate(trend = case_when(
          value >= support_threshold ~ "Positive",
          value <= negsupport_threshold ~ "Negative",
          TRUE ~ "Neutral")) 

post_table_red <- post_estimates_red %>%
    mutate(genome=factor(genome, levels=rev(hmsc_tree_red$tip.label))) %>%
    select(-trend) %>%
    mutate(value = case_when(
          value >= support_threshold ~ "Positive",
          value <= negsupport_threshold ~ "Negative",
          TRUE ~ "Neutral")) %>%
    mutate(value=factor(value, levels=c("Positive","Neutral","Negative"))) %>%
    pivot_wider(names_from = variable, values_from = value) %>%
    rename(intercept=2,
           male=3,
           index500=4,
           autumn=5,
           winter=6,
           logseqdepth=7) %>%
   column_to_rownames(var="genome") %>%
  select(-intercept)

# Posterior estimates grey
post_estimates_grey <- getPostEstimate(hM=m.grey, parName="Beta")$support %>%
    as.data.frame() %>%
    mutate(variable=m.grey$covNames) %>%
    pivot_longer(!variable, names_to = "genome", values_to = "value") %>%
    mutate(trend = case_when(
          value >= support_threshold ~ "Positive",
          value <= negsupport_threshold ~ "Negative",
          TRUE ~ "Neutral")) 

post_table_grey <- post_estimates_grey %>%
    mutate(genome=factor(genome, levels=rev(hmsc_tree_grey$tip.label))) %>%
    select(-trend) %>%
    mutate(value = case_when(
          value >= support_threshold ~ "Positive",
          value <= negsupport_threshold ~ "Negative",
          TRUE ~ "Neutral")) %>%
    mutate(value=factor(value, levels=c("Positive","Neutral","Negative"))) %>%
    pivot_wider(names_from = variable, values_from = value) %>%
    rename(intercept=2,
           male=3,
           index500=4,
           autumn=5,
           winter=6,
           logseqdepth=7) %>%
   column_to_rownames(var="genome") %>%
  select(-intercept) 
  
# Basal ggtree
postestimates_tree <- hmsc_tree_all %>%
        force.ultrametric(.,method="extend") %>%
        ggtree(., size = 0.3)
        
#Add phylum colors next to the tree tips
postestimates_tree <- gheatmap(postestimates_tree, phylums, offset=-0.2, width=0.1, colnames=FALSE) +
      scale_fill_manual(values=custom_colors)+
      labs(fill="Phylum")

#Reset fill scale to use a different colour profile in the heatmap
postestimates_tree <- postestimates_tree + new_scale_fill()

# Add posterior significant heatmap for red
postestimates_tree_red <- gheatmap(postestimates_tree, post_table_red, offset=0.15, width=0.5, colnames=TRUE, colnames_position="top",colnames_angle=90, colnames_offset_y=1, hjust=0) 

# Add posterior significant heatmap for grey
postestimates_tree_both <- gheatmap(postestimates_tree_red, post_table_grey, offset=1.5, width=0.5, colnames=TRUE, colnames_position="top",colnames_angle=90, colnames_offset_y=1, hjust=0) +
        scale_fill_manual(values=c("#be3e2b","#BABABA","#b2b530"), na.value="white")+
        labs(fill="Trend") 


postestimates_tree_both +
        vexpand(.15, 1) + # expand top 
        vexpand(.05, -1) + # expand bottom
        annotate('text', x=3.25, y=-8, label = "S. vulgaris", color='black', fontface=2) +
        annotate('text', x=4.6, y=-8, label = "S. carolinensis", color='black', fontface=2)  

```


## Correlations among bacteria

### Red squirrel

```{r hmsc_corr_red, warning=FALSE, comments="", message=FALSE, fig.dim=c(8,8), cache=TRUE}
#Compute the residual correlation matrix
OmegaCor = computeAssociations(m.red)

#Co-occurrence matrix at the animal level
supportLevel = 0.95
toPlot = ((OmegaCor[[1]]$support>supportLevel)
          + (OmegaCor[[1]]$support<(1-supportLevel))>0)*OmegaCor[[1]]$mean

matrix <- toPlot %>% 
      as.data.frame() %>%
      rownames_to_column(var="genome1") %>%
      pivot_longer(!genome1, names_to = "genome2", values_to = "cor") %>%
      mutate(genome1= factor(genome1, levels=rev(hmsc_tree_red$tip.label))) %>%
      mutate(genome2= factor(genome2, levels=hmsc_tree_red$tip.label)) %>%
      ggplot(aes(x = genome1, y = genome2, fill = cor)) +
            geom_tile() + 
            scale_fill_gradient2(low = "#be3e2b",
                       mid = "#f4f4f4",
                       high = "#b2b530")+
            theme_void() +
            theme(legend.position = "none") 

corr.legend <- get_legend(matrix, position="right") 
corr.legend <- as_ggplot(corr.legend)

vtree <- hmsc_tree_red %>%
  force.ultrametric(.,method="extend") %>%
  ggtree(., expand=1.5) +
  hexpand(0.5) 

#Add phylum colors next to the tree tips
vtree <- gheatmap(vtree, phylums, offset=-0.1, width=0.6, colnames=FALSE) +
      scale_fill_manual(values=custom_colors) +
      theme(legend.position = 'none') + scale_y_reverse()

vtreeD <- hmsc_tree_red %>%
  force.ultrametric(.,method="extend") %>%
  ggtree(., expand=1.5, layout="dendrogram") 

#Add phylum colors next to the tree tips
vtreeD <- gheatmap(vtreeD, phylums, offset=-0.1, width=0.6, colnames=FALSE) +
      scale_fill_manual(values=custom_colors) +
      theme(legend.position = 'none') 

# properly align trees to corr matrix with package aplot
matrix %>%  insert_left(vtree, width=.25) %>% insert_top(vtreeD, height=.3) %>% insert_right(corr.legend, width=0.15)

```

### Grey squirrel

```{r hmsc_corr_grey, warning=FALSE, comments="", message=FALSE, fig.dim=c(8,8), cache=TRUE}
#Compute the residual correlation matrix
OmegaCor = computeAssociations(m.grey)

#Co-occurrence matrix at the animal level
supportLevel = 0.95
toPlot = ((OmegaCor[[1]]$support>supportLevel)
          + (OmegaCor[[1]]$support<(1-supportLevel))>0)*OmegaCor[[1]]$mean

matrix <- toPlot %>% 
      as.data.frame() %>%
      rownames_to_column(var="genome1") %>%
      pivot_longer(!genome1, names_to = "genome2", values_to = "cor") %>%
      mutate(genome1= factor(genome1, levels=rev(hmsc_tree_grey$tip.label))) %>%
      mutate(genome2= factor(genome2, levels=hmsc_tree_grey$tip.label)) %>%
      ggplot(aes(x = genome1, y = genome2, fill = cor)) +
            geom_tile() + 
            scale_fill_gradient2(low = "#be3e2b",
                       mid = "#f4f4f4",
                       high = "#b2b530")+
            theme_void() +
            theme(legend.position = "none") 

corr.legend <- get_legend(matrix, position="right") 
corr.legend <- as_ggplot(corr.legend)

vtree <- hmsc_tree_grey %>%
  force.ultrametric(.,method="extend") %>%
  ggtree(., expand=1.5) +
  hexpand(0.5) 

#Add phylum colors next to the tree tips
vtree <- gheatmap(vtree, phylums, offset=-0.1, width=0.6, colnames=FALSE) +
      scale_fill_manual(values=custom_colors) +
      theme(legend.position = 'none') + scale_y_reverse()

vtreeD <- hmsc_tree_grey %>%
  force.ultrametric(.,method="extend") %>%
  ggtree(., expand=1.5, layout="dendrogram") 

#Add phylum colors next to the tree tips
vtreeD <- gheatmap(vtreeD, phylums, offset=-0.1, width=0.6, colnames=FALSE) +
      scale_fill_manual(values=custom_colors) +
      theme(legend.position = 'none') 

# properly align trees to corr matrix with package aplot
matrix %>%  insert_left(vtree, width=.25) %>% insert_top(vtreeD, height=.3) %>% insert_right(corr.legend, width=0.15)

```


## Predictions by urbanisation


### Selection of predictive MAGs

```{r pred_mags}

pred_mags.red<- var_pred_table.red%>%
  filter(var_pred>=0.001) %>%
  pull(mag)

pred.ab.red <- genome_counts %>%
  mutate_at(vars(-genome),~./sum(.)) %>% #apply TSS normalisation
  filter(genome %in% m.red$spNames) %>%
   rowwise() %>% #compute for each row (genome)
   mutate(
    mean = mean(c_across(all_of(red_samples)), na.rm = TRUE),  # Get mean genome counts across red samples
    prev = sum(c_across(all_of(red_samples)) > 0, na.rm = TRUE) / sum(!is.na(c_across(all_of(red_samples)))), # Proportion of non-zero values
   subset = ifelse(genome %in% pred_mags.red, 'pred', 'nonpred')) %>%
   dplyr::select(genome, subset, mean, prev) %>%
   left_join(genome_metadata, by=join_by(genome==genome)) %>%
   arrange(subset,-mean)

pred.plot.red <- pred.ab.red %>%
  mutate(phylum = ifelse(subset=='pred',phylum, NA)) %>%
  ggplot(., aes(x=prev, y=mean, color=phylum)) +
  geom_point(size=3) +
  scale_color_manual(values=custom_colors) +
  theme_bw()+
    labs(x="Prevalence", y="Mean abundance")

pred_mags.grey<- var_pred_table.grey%>%
  filter(var_pred>=0.001) %>%
  pull(mag)

pred.ab.grey <- genome_counts %>% 
  mutate_at(vars(-genome),~./sum(.)) %>% #apply TSS normalisation
  filter(genome %in% m.grey$spNames) %>%
   rowwise() %>% #compute for each row (genome)
   mutate(
    mean = mean(c_across(all_of(grey_samples)), na.rm = TRUE),  # Get mean genome counts across red samples
    prev = sum(c_across(all_of(grey_samples)) > 0, na.rm = TRUE) / sum(!is.na(c_across(all_of(grey_samples)))), # Proportion of non-zero values
    subset = ifelse(genome %in% pred_mags.grey, 'pred', 'nonpred')) %>%
   dplyr::select(genome, subset, mean, prev) %>%
   left_join(genome_metadata, by=join_by(genome==genome)) %>%
   arrange(subset,-mean)

pred.plot.grey <- pred.ab.grey %>%
  mutate(phylum = ifelse(subset=='pred',phylum, NA)) %>%
  ggplot(., aes(x=prev, y=mean, color=phylum)) +
  geom_point(size=3) +
  scale_color_manual(values=custom_colors) +
  theme_bw()+
    labs(x="Prevalence", y="Mean abundance") +
    theme(y.axis.title =element_blank())

ggarrange(pred.plot.red, pred.plot.grey, legend="right", common.legend = TRUE,
          ncol = 2, nrow=1)


```


### Compositional predictions

```{r hmsc_pred_urb_red, warning=FALSE, message=FALSE}

gradient = seq(0,1, by=0.1)
gradientlength = length(gradient)

pred_urban_red <- constructGradient(m.red,
                      focalVariable = "index500",
                      non.focalVariables = list(logseqdepth=list(1),species=list(1), season=list(1), sex=list(1)),
                      ngrid=gradientlength) %>%
                      predict(m.red, Gradient = ., expected = TRUE) %>%
                      do.call(rbind,.) %>%
                      as.data.frame() %>%
                      mutate(index500=rep(gradient,1000)) %>%
                      pivot_longer(-c(index500), names_to = "genome", values_to = "value")

urbcomp_red <- post_estimates_red %>%
    filter(variable=="index500",
           genome %in% pred_mags.red) %>% #keep only mags predictive mags
    select(genome,trend) %>%
    left_join(pred_urban_red, by=join_by(genome==genome)) %>%
    group_by(genome, trend, index500) %>%
    summarize(value = mean(value, na.rm = TRUE)) %>%
    left_join(genome_metadata, by=join_by(genome == genome))

urbcomp_red %>%
  filter(trend !="Neutral") %>%
  group_by(genome) %>%
  mutate(slope = value[index500 == 1] - value[index500 == 0]) %>%
  left_join(genome_metadata) %>%
  distinct(genome, .keep_all=TRUE) %>%
  select(-c(3,4,12,13,14)) %>%
  relocate(slope, .after=trend) %>%
  group_by(genus) %>%
  summarise(mean_slope=mean(slope),
            sd_slope=sd(slope)) %>%
  arrange(-mean_slope) %>%
  paged_table()

urbcomp_plot_red <- urbcomp_red %>%
    ggplot(aes(x=index500, y=value, group=genome, color=phylum, linetype=trend)) +
        geom_line() +
        scale_linetype_manual(values=c("solid","dashed","solid")) +
        scale_color_manual(values=custom_colors) +
        facet_grid(fct_rev(trend) ~ phylum) +
        labs(y="Genome abundance (log)",x="Urbanization index") +
        theme(legend.position = "none") +
        theme_minimal() +
         theme(legend.position = "none",
               axis.text.x = element_text(angle = 45, hjust = 0.8,),
               axis.line.x = element_line(size = 0.5, linetype = "solid", colour = "black"),
               strip.text.x = element_text(angle = 90, hjust = 0,),
               strip.text.y = element_text(face="bold"),
)

taxonomy_urban_red <- post_estimates_red %>%
    filter(variable=="index500", 
           genome %in% pred_mags.red,
           trend =="Positive") %>%
    pull(genome)

taxonomy_rural_red <- post_estimates_red %>%
    filter(variable=="index500", 
           genome %in% pred_mags.red,
           trend =="Negative") %>%
    pull(genome)

```


```{r hmsc_pred_urb_grey, fig.dim=c(10,10), warning=FALSE, message=FALSE}

gradient = seq(0,1, by=0.1)
gradientlength = length(gradient)

pred_urban_grey <- constructGradient(m.grey,
                      focalVariable = "index500",
                      non.focalVariables = list(logseqdepth=list(1),species=list(1), season=list(1), sex=list(1)),
                      ngrid=gradientlength) %>%
                      predict(m.grey, Gradient = ., expected = TRUE) %>%
                      do.call(rbind,.) %>%
                      as.data.frame() %>%
                      mutate(index500=rep(gradient,1000)) %>%
                      pivot_longer(-c(index500), names_to = "genome", values_to = "value")

urbcomp_grey <- post_estimates_grey %>%
    filter(variable=="index500",
           genome %in% pred_mags.grey) %>% #keep only mags predictive mags
    select(genome,trend) %>%
    left_join(pred_urban_grey, by=join_by(genome==genome)) %>%
    group_by(genome, trend, index500) %>%
    summarize(value = mean(value, na.rm = TRUE)) %>%
    left_join(genome_metadata, by=join_by(genome == genome))

urbcomp_grey %>%
  filter(trend !="Neutral") %>%
  group_by(genome) %>%
  mutate(slope = value[index500 == 1] - value[index500 == 0]) %>%
  left_join(genome_metadata) %>%
  distinct(genome, .keep_all=TRUE) %>%
  select(-c(3,4,12,13,14)) %>%
  relocate(slope, .after=trend) %>%
  group_by(genus) %>%
  summarise(mean_slope=mean(slope),
            sd_slope=sd(slope)) %>%
  arrange(-mean_slope) %>%
  paged_table()
  

urbcomp_plot_grey <- urbcomp_grey %>%
    ggplot(aes(x=index500, y=value, group=genome, color=phylum, linetype=trend)) +
        geom_line() +
        scale_linetype_manual(values=c("solid","dashed","solid")) +
        scale_color_manual(values=custom_colors) +
        facet_grid(fct_rev(trend) ~ phylum) +
        labs(y="Genome abundance (log)",x="Urbanization index") +
        theme(legend.position = "none") +
        theme_minimal() +
         theme(legend.position = "none",
               axis.text.x = element_text(angle = 45, hjust = 0.8,),
               axis.line.x = element_line(size = 0.5, linetype = "solid", colour = "black"),
               strip.text.x = element_text(angle = 90, hjust = 0,),
               strip.text.y = element_text(face="bold"))

taxonomy_urban_grey <- post_estimates_grey %>%
    filter(variable=="index500", 
           genome %in% pred_mags.grey,
           trend =="Positive") %>%
    pull(genome)

taxonomy_rural_grey <- post_estimates_grey %>%
    filter(variable=="index500", 
           genome %in% pred_mags.grey,
           trend =="Negative") %>%
    pull(genome)

ggarrange(urbcomp_plot_red, urbcomp_plot_grey, ncol=1, nrow=2)

```

### Functional predictions

```{r GIFTs_data_preparation2, echo = T, results = 'hide', warning=FALSE, message=FALSE}

tss <- function(abund){sweep(abund, 2, colSums(abund), FUN="/")} 

genome_counts2 <- genome_counts %>%
  column_to_rownames(var="genome")

#Get list of present MAGs
present_MAGs <- genome_counts2 %>%
  filter(rowSums(.[, -1]) != 0) %>%
  rownames()

#Align distillr annotations with present MAGs and remove all-zero and all-one traits
present_MAGs <- present_MAGs[present_MAGs %in% rownames(genome_gifts)]
genome_gifts_filt <- genome_gifts[present_MAGs,] %>%
  select_if(~!all(. == 0)) %>%  #remove all-zero modules
  select_if(~!all(. == 1)) #remove all-one modules

GIFTs_elements <- to.elements(genome_gifts_filt, GIFT_db2)

#Aggregate element-level GIFTs into the function level
GIFTs_functions <- to.functions(GIFTs_elements, GIFT_db2)

#Aggregate function-level GIFTs into overall Biosynthesis, Degradation and Structural GIFTs and get overall metabolic capacity indices per MAG (at the domain level)
GIFTs_domains <- to.domains(GIFTs_functions, GIFT_db2) %>% as.data.frame() %>%
  mutate(Overall=rowMeans(select(.,Biosynthesis,Structure,Degradation), na.rm=TRUE))

```

#### Functions

```{r function_urb_pred_red, results="hide", eval=TRUE, message=FALSE, warning=FALSE, cache=TRUE}

community_func_urb_red <- pred_urban_red %>%
  filter(genome %in% pred_mags.red) %>% #keep only predictive mags
  group_by(index500, genome) %>%
  mutate(row_id = row_number()) %>%
  pivot_wider(names_from = genome, values_from = value) %>%
  ungroup() %>%
  group_split(row_id) %>%
  as.list() %>%
  lapply(., FUN = function(x){x %>%
    select(-row_id) %>%
    column_to_rownames(var = "index500") %>%
    as.data.frame() %>%
    exp() %>%
    t() %>%
    tss() %>%
    to.community(GIFTs_functions,.,GIFT_db2) %>%
    as.data.frame() %>%
    rownames_to_column(var="index500")
   })


calculate_slope <- function(x) {
  lm_fit <- lm(unlist(x) ~ seq_along(unlist(x)))
  coef(lm_fit)[2]
}

function_pred_urb_red <- map_dfc(community_func_urb_red, function(mat) {
      mat %>%
        column_to_rownames(var = "index500") %>%
        t() %>%
        as.data.frame() %>%
        rowwise() %>%
        mutate(slope = calculate_slope(c_across(everything()))) %>%
        select(slope) }) %>%
      t() %>%
      as.data.frame() %>%
      set_names(colnames(community_func_urb_red[[1]])[-1]) %>%
      rownames_to_column(var="iteration") %>%
      pivot_longer(!iteration, names_to="trait",values_to="value") %>%
      group_by(trait) %>%
      summarise(mean=mean(value),
        p1 = quantile(value, probs = 0.1),
        p9 = quantile(value, probs = 0.9),
        positive_support = sum(value > 0)/1000,
        negative_support = sum(value < 0)/1000) %>%
      arrange(-positive_support)

```


```{r function_urb_plot_red, message=FALSE, warning=FALSE, fig.dim=c(10, 8)}

# Positively associated with urbanisation
function_pred_urb_red %>%
  filter(mean >0, positive_support>=0.9) %>%
  arrange(-mean) %>%
  select(-negative_support) %>%
  rename(support=positive_support) %>%
  left_join(GIFT_db2 %>% select(Code_function, Function),by=join_by(trait==Code_function)) %>%
  relocate(Function, .after=trait) %>%
  unique() %>%
  paged_table()

#Negatively associated with urbanisation
function_pred_urb_red %>%
  filter(mean <0, negative_support>=0.9) %>%
  arrange(-mean) %>%
  select(-positive_support) %>%
  rename(support=negative_support) %>%
  left_join(GIFT_db2 %>% select(Code_function, Function),by=join_by(trait==Code_function)) %>%
  relocate(Function, .after=trait) %>%
  unique() %>%
  paged_table()
  

all_functions_red <- #bind_rows(positive_red,negative_red)  %>%
  function_pred_urb_red %>%
  left_join(GIFT_db2,by=join_by(trait==Code_function)) %>%
  mutate(trait=factor(trait)) %>%
  mutate(function_legend=str_c(trait," - ",Function)) %>%
  select(trait,mean,p1,p9,function_legend) %>% 
  unique()

gift_colors <- read_tsv("data/gift_colors.tsv") 

red_fun_urb_plot <- all_functions_red %>%
  ggplot(aes(x=mean, y=function_legend, xmin=p1, xmax=p9, color=function_legend)) +
      geom_point() +
      geom_errorbar() +
      xlim(c(-0.12,0.12)) +
      geom_vline(xintercept=0) +
      scale_color_manual(values = gift_colors$Color) +
      theme_minimal() +
      labs(x="Regression coefficient",y="Functional trait", title= "S. vulgaris") +
      #guides(col = guide_legend(ncol = 1))
      theme(legend.position = "none",
             plot.title=element_text( hjust=0.5, vjust=0.6, face='bold', size=12))

```



```{r function_urb_pred_grey, results="hide", eval=TRUE, message=FALSE, warning=FALSE, cache=TRUE}

community_func_urb_grey  <- pred_urban_grey %>%
  filter(genome %in% pred_mags.grey) %>% #keep only predictive mags
  group_by(index500, genome) %>%
  mutate(row_id = row_number()) %>%
  pivot_wider(names_from = genome, values_from = value) %>%
  ungroup() %>%
  group_split(row_id) %>%
  as.list() %>%
  lapply(., FUN = function(x){x %>%
    select(-row_id) %>%
    column_to_rownames(var = "index500") %>%
    as.data.frame() %>%
    exp() %>%
    t() %>%
    tss() %>%
    to.community(GIFTs_functions,.,GIFT_db2) %>%
    as.data.frame() %>%
    rownames_to_column(var="index500")
   })


calculate_slope <- function(x) {
  lm_fit <- lm(unlist(x) ~ seq_along(unlist(x)))
  coef(lm_fit)[2]
}

function_pred_urb_grey <- map_dfc(community_func_urb_grey, function(mat) {
      mat %>%
        column_to_rownames(var = "index500") %>%
        t() %>%
        as.data.frame() %>%
        rowwise() %>%
        mutate(slope = calculate_slope(c_across(everything()))) %>%
        select(slope) }) %>%
      t() %>%
      as.data.frame() %>%
      set_names(colnames(community_func_urb_grey[[1]])[-1]) %>%
      rownames_to_column(var="iteration") %>%
      pivot_longer(!iteration, names_to="trait",values_to="value") %>%
      group_by(trait) %>%
      summarise(mean=mean(value),
        p1 = quantile(value, probs = 0.1),
        p9 = quantile(value, probs = 0.9),
        positive_support = sum(value > 0)/1000,
        negative_support = sum(value < 0)/1000) %>%
      arrange(-positive_support)


```


```{r function_urb_plot_grey, message=FALSE, warning=FALSE, fig.dim=c(10, 8)}

# Positively associated with urbanisation
function_pred_urb_grey %>%
  filter(mean >0, positive_support>=0.9) %>%
  arrange(-mean) %>%
  select(-negative_support) %>%
  rename(support=positive_support) %>%
  left_join(GIFT_db2 %>% select(Code_function, Function),by=join_by(trait==Code_function)) %>%
  relocate(Function, .after=trait) %>%
  unique() %>%
  paged_table()

#Negatively associated with urbanisation
function_pred_urb_grey %>%
  filter(mean <0, negative_support>=0.9) %>%
  arrange(-mean) %>%
  select(-positive_support) %>%
  rename(support=negative_support) %>%
  left_join(GIFT_db2 %>% select(Code_function, Function),by=join_by(trait==Code_function)) %>%
  relocate(Function, .after=trait) %>%
  unique() %>%
  paged_table()


all_functions_grey <- #bind_rows(positive_grey,negative_grey) %>%
  function_pred_urb_grey %>%
  left_join(GIFT_db2,by=join_by(trait==Code_function)) %>%
  mutate(trait=factor(trait)) %>%
  mutate(function_legend=str_c(trait," - ",Function)) %>%
  select(trait,mean,p1,p9,function_legend) %>% 
  unique()

grey_fun_urb_plot <- all_functions_grey %>%
  ggplot(aes(x=mean, y=function_legend, xmin=p1, xmax=p9, color=function_legend)) +
      geom_point() +
      geom_errorbar() +
      xlim(-0.12,0.12) +
      geom_vline(xintercept=0) + 
      scale_color_manual(values = gift_colors$Color) +
      theme_minimal() +
      labs(x="Regression coefficient",y="Functional trait", title="S. carolinensis") +
      #guides(col = guide_legend(ncol = 1))
      theme(legend.position = "none",
            axis.title.y = element_blank(),
            axis.text.y = element_blank(),
            plot.title=element_text( hjust=0.5, vjust=0.6, face='bold', size=12),
            plot.margin = margin(l = 40))
      

      # plot.margin = margin(t = 40)) 
      # annotate('text', x=0, y=21, label = "S. carolinensis", color='black', fontface=2)


```


```{r fun_urban_plot, message=FALSE, warning=FALSE, fig.fullwidth=TRUE}

ggarrange(red_fun_urb_plot, grey_fun_urb_plot,
          ncol=2, nrow=1)

```


#### Elements


```{r element_urb_pred_red, results="hide", eval=TRUE, message=FALSE, warning=FALSE, cache=TRUE}

community_elem_urb_red <-  pred_urban_red %>%
  filter(genome %in% pred_mags.red) %>% #keep only predictive mags
  group_by(index500, genome) %>%
  mutate(row_id = row_number()) %>%
  pivot_wider(names_from = genome, values_from = value) %>%
  ungroup() %>%
  group_split(row_id) %>%
  as.list() %>%
  lapply(., FUN = function(x){x %>%
    select(-row_id) %>%
    column_to_rownames(var = "index500") %>%
    as.data.frame() %>%
    exp() %>%
    t() %>%
    tss() %>%
    to.community(GIFTs_elements,.,GIFT_db2) %>%
    as.data.frame() %>%
    rownames_to_column(var="index500")
   })


calculate_slope <- function(x) {
  lm_fit <- lm(unlist(x) ~ seq_along(unlist(x)))
  coef(lm_fit)[2]
}

element_pred_urb_red <- map_dfc(community_elem_urb_red, function(mat) {
      mat %>%
        column_to_rownames(var = "index500") %>%
        t() %>%
        as.data.frame() %>%
        rowwise() %>%
        mutate(slope = calculate_slope(c_across(everything()))) %>%
        select(slope) }) %>%
      t() %>%
      as.data.frame() %>%
      set_names(colnames(community_elem_urb_red[[1]])[-1]) %>%
      rownames_to_column(var="iteration") %>%
      pivot_longer(!iteration, names_to="trait",values_to="value") %>%
      group_by(trait) %>%
      summarise(mean=mean(value),
        p1 = quantile(value, probs = 0.1),
        p9 = quantile(value, probs = 0.9),
        positive_support = sum(value > 0)/1000,
        negative_support = sum(value < 0)/1000) %>%
      arrange(-positive_support)

```


```{r element_urb_plot_red, message=FALSE, warning=FALSE, fig.dim=c(10, 8)}

# Positively associated with urbanisation
element_pred_urb_red %>%
  filter(mean >0, positive_support>=0.9) %>%
  arrange(-mean) %>%
  select(-negative_support) %>%
  rename(support=positive_support) %>%
  left_join(GIFT_db2 %>% select(Code_element, Element, Function),by=join_by(trait==Code_element)) %>%
  relocate(Element, Function, .after=trait) %>%
  unique() %>%
  paged_table()

#Negatively associated with urbanisation
element_pred_urb_red %>%
  filter(mean <0, negative_support>=0.9) %>%
  arrange(mean) %>%
  select(-positive_support) %>%
  rename(support=negative_support) %>%
  left_join(GIFT_db2 %>% select(Code_element, Element, Function),by=join_by(trait==Code_element)) %>%
   relocate(Element, Function, .after=trait) %>%
  unique() %>%
  paged_table()


positive_urb_red <- element_pred_urb_red%>%
  filter(mean >0) %>%
  arrange(mean) %>%
  filter(positive_support>=0.9) %>%
  select(-negative_support) %>%
  rename(support=positive_support)
negative_urb_red <- element_pred_urb_red %>%
  filter(mean <0) %>%
  arrange(mean) %>%
  filter(negative_support>=0.9) %>%
  select(-positive_support) %>%
  rename(support=negative_support)

all_elements_urb_red <- bind_rows(positive_urb_red,negative_urb_red) %>%
  left_join(GIFT_db2,by=join_by(trait==Code_element)) %>%
  mutate(trait=factor(trait,levels=c(rev(positive_urb_red$trait),rev(negative_urb_red$trait)))) %>%
  mutate(Code_function=factor(Code_function)) %>%
  mutate(element_legend=str_c(trait," - ",Element)) %>%
  mutate(function_legend=str_c(Code_function," - ",Function)) %>%
  select(trait,mean,p1,p9,element_legend,function_legend) %>% 
  unique()

red_elem_urb_plot <- all_elements_urb_red %>%
  ggplot(aes(x=mean, y=fct_reorder(element_legend, mean), xmin=p1, xmax=p9, color=function_legend)) +
      geom_point() +
      geom_errorbar() +
      #xlim(c(-0.050,0.050)) +
      geom_vline(xintercept=0) +
      scale_color_manual(values = gift_colors$Color) +
      theme_minimal() +
      labs(x="Regression coefficient",y="Functional trait", title="S. vulgaris", color="Function") +
      #guides(col = guide_legend(ncol = 1))
      theme(legend.position = "right",
            plot.title=element_text( hjust=0.45, vjust=0.6, face='bold', size=12))

```


```{r element_urb_pred_grey, results="hide", eval=TRUE, message=FALSE, warning=FALSE, cache=TRUE}

community_elem_urb_grey <- pred_urban_grey  %>%
  filter(genome %in% pred_mags.grey) %>% #keep only predictive mags
  group_by(index500, genome) %>%
  mutate(row_id = row_number()) %>%
  pivot_wider(names_from = genome, values_from = value) %>%
  ungroup() %>%
  group_split(row_id) %>%
  as.list() %>%
  lapply(., FUN = function(x){x %>%
    select(-row_id) %>%
    column_to_rownames(var = "index500") %>%
    as.data.frame() %>%
    exp() %>%
    t() %>%
    tss() %>%
    to.community(GIFTs_elements,.,GIFT_db2) %>%
    as.data.frame() %>%
    rownames_to_column(var="index500")
   })


calculate_slope <- function(x) {
  lm_fit <- lm(unlist(x) ~ seq_along(unlist(x)))
  coef(lm_fit)[2]
}

element_pred_urb_grey <- map_dfc(community_elem_urb_grey, function(mat) {
      mat %>%
        column_to_rownames(var = "index500") %>%
        t() %>%
        as.data.frame() %>%
        rowwise() %>%
        mutate(slope = calculate_slope(c_across(everything()))) %>%
        select(slope) }) %>%
      t() %>%
      as.data.frame() %>%
      set_names(colnames(community_elem_urb_grey[[1]])[-1]) %>%
      rownames_to_column(var="iteration") %>%
      pivot_longer(!iteration, names_to="trait",values_to="value") %>%
      group_by(trait) %>%
      summarise(mean=mean(value),
        p1 = quantile(value, probs = 0.1),
        p9 = quantile(value, probs = 0.9),
        positive_support = sum(value > 0)/1000,
        negative_support = sum(value < 0)/1000) %>%
      arrange(-positive_support)

```


```{r element_urb_plot_grey, message=FALSE, warning=FALSE, fig.dim=c(10, 8)}

# Positively associated with urbanisation
element_pred_urb_grey %>%
  filter(mean >0, positive_support>=0.9) %>%
  arrange(-mean) %>%
  select(-negative_support) %>%
  rename(support=positive_support) %>%
  left_join(GIFT_db2 %>% select(Code_element, Element, Function),by=join_by(trait==Code_element)) %>%
  relocate(Element, Function, .after=trait) %>%
  unique() %>%
  paged_table()

#Negatively associated with urbanisation
element_pred_urb_grey %>%
  filter(mean <0, negative_support>=0.9) %>%
  arrange(mean) %>%
  select(-positive_support) %>%
  rename(support=negative_support) %>%
  left_join(GIFT_db2 %>% select(Code_element, Element, Function),by=join_by(trait==Code_element)) %>%
  relocate(Element, Function, .after=trait) %>%
  unique() %>%
  paged_table()

positive_urb_grey <- element_pred_urb_grey %>%
  filter(mean >0) %>%
  arrange(mean) %>%
  filter(positive_support>=0.9) %>%
  select(-negative_support) %>%
  rename(support=positive_support)
negative_urb_grey <- element_pred_urb_grey %>%
  filter(mean <0) %>%
  arrange(mean) %>%
  filter(negative_support>=0.9) %>%
  select(-positive_support) %>%
  rename(support=negative_support)


all_elements_urb_grey <- bind_rows(positive_urb_grey,negative_urb_grey) %>%
  left_join(GIFT_db2,by=join_by(trait==Code_element)) %>%
  mutate(trait=factor(trait,levels=c(rev(positive_urb_grey$trait),rev(negative_urb_grey$trait)))) %>%
  mutate(Code_function=factor(Code_function)) %>%
  mutate(element_legend=str_c(trait," - ",Element)) %>%
  mutate(function_legend=str_c(Code_function," - ",Function)) %>%
  select(trait,mean,p1,p9,element_legend,function_legend) %>% 
  unique()

grey_elem_urb_plot <- all_elements_urb_grey %>%
  ggplot(aes(x=mean, y=fct_reorder(element_legend, mean), xmin=p1, xmax=p9, color=function_legend)) +
      geom_point() +
      geom_errorbar() +
      #xlim(c(-0.050,0.050)) +
      geom_vline(xintercept=0) +
      scale_color_manual(values = gift_colors$Color) +
      theme_minimal() +
      labs(x="Regression coefficient",y="Functional trait", title="S. carolinensis", color="Function") +
      guides(col = guide_legend(ncol = 5)) +
      theme(legend.position = "right",
            axis.title.y = element_blank(),
            plot.title=element_text( hjust=0.5, vjust=0.6, face='bold', size=12))

```

```{r elem_urban_plot, message=FALSE, warning=FALSE, fig.dim=c(10,10)}

el.legend <- get_legend(grey_elem_urb_plot)

ggarrange(red_elem_urb_plot, grey_elem_urb_plot, legend.grob=el.legend, legend="bottom", common.legend=TRUE, 
          ncol=2, nrow=1)

```



### Sources of functional variation


```{r gift_pcoa, warning=FALSE, message=FALSE}

gift_pcoa <- genome_gifts %>%
    to.elements(., GIFT_db2) %>%
    as.data.frame() %>%
    vegdist(method="euclidean") %>%
    pcoa()

gift_pcoa_rel_eigen <- gift_pcoa$values$Relative_eig[1:10]


# Get genome positions
gift_pcoa_vectors <- gift_pcoa$vectors %>% #extract vectors
  as.data.frame() %>% 
  select(Axis.1,Axis.2) # keep the first 2 axes

gift_pcoa_eigenvalues <- gift_pcoa$values$Eigenvalues[c(1,2)]


```

```{r ord_urb_red, warning=FALSE, message=FALSE}

gift_pcoa_urb_red <- cov(genome_gifts, scale(gift_pcoa_vectors)) %*% diag((gift_pcoa_eigenvalues/(nrow(genome_gifts)-1))^(-0.5)) %>%
  as.data.frame() %>% 
  rename(Axis.1=1,Axis.2=2) %>% 
  rownames_to_column(var="label") %>% 
  #get function summary vectors
  mutate(func=substr(label,1,5)) %>% 
  group_by(func) %>% 
  summarise(Axis.1=mean(Axis.1),
            Axis.2=mean(Axis.2)) %>% 
  rename(label=func) %>% 
  filter(label %in% (positive_urb_red %>% 
    pull(trait)))


scale <- 15 # scale for vector loadings

gift_pcoa_vectors %>% 
  rownames_to_column(var="genome") %>% 
  left_join(genome_metadata, by="genome") %>%
  mutate(relation = case_when(genome %in% taxonomy_urban_red ~ "urban",
                              genome %in% taxonomy_rural_red ~ "rural",
                              TRUE ~ "neutral")) %>%
 group_by(relation) %>%
  mutate(x_cen = mean(Axis.1, na.rm = TRUE)) %>%
  mutate(y_cen = mean(Axis.2, na.rm = TRUE)) %>%
  ungroup() %>% 
  mutate(phylum=ifelse(relation=="neutral","NA",phylum)) %>% 
  arrange(phylum)%>%
  ggplot() +
      #genome positions
      scale_color_manual(values=custom_colors,na.value = "#EDEDED")+
      geom_point(aes(x=Axis.1,y=Axis.2, color=phylum, size=length), alpha=0.9, shape=16) +
      new_scale_color() +
      scale_color_manual(values = c("urban" = "#b592ad", "rural" = "#849c20")) + 
      geom_segment(data = . %>% filter(relation != "neutral"),
                   aes(x = x_cen, y = y_cen, xend = Axis.1, yend = Axis.2, color=relation), 
                   alpha = 0.5) +      
      new_scale_color() +
      #loading positions
      geom_segment(data=gift_pcoa_urb_red, 
                   aes(x=0, y=0, xend=Axis.1 * scale, yend=Axis.2 * scale),
                    arrow = arrow(length = unit(0.3, "cm"), 
                    type = "open", 
                    angle = 25),
                    linewidth = 0.5, 
                    color = "black") +
     #Primary and secondary scale adjustments
     scale_x_continuous(name = paste0("PCoA1 (",round(gift_pcoa_rel_eigen[1]*100, digits = 2), " %)"),
                      sec.axis = sec_axis(~ . / scale, name = "Loadings on PCoA1")
            ) +
     scale_y_continuous(name = paste0("PCoA2 (",round(gift_pcoa_rel_eigen[2]*100, digits = 2), " %)"),
                      sec.axis = sec_axis(~ . / scale, name = "Loadings on PCoA2")
            ) +
    geom_label_repel(data = gift_pcoa_urb_red,
                     aes(label = label, x = Axis.1 * scale, y = Axis.2 * scale),
                     segment.color = 'transparent') +
    theme_minimal() + 
    theme(legend.position = "none")

```

```{r ord_urb_grey, warning=FALSE, message=FALSE}

gift_pcoa_urb_grey <- cov(genome_gifts, scale(gift_pcoa_vectors)) %*% diag((gift_pcoa_eigenvalues/(nrow(genome_gifts)-1))^(-0.5)) %>%
  as.data.frame() %>% 
  rename(Axis.1=1,Axis.2=2) %>% 
  rownames_to_column(var="label") %>% 
  #get function summary vectors
  mutate(func=substr(label,1,5)) %>% 
  group_by(func) %>% 
  summarise(Axis.1=mean(Axis.1),
            Axis.2=mean(Axis.2)) %>% 
  rename(label=func) %>% 
  filter(label %in% (positive_urb_grey %>% 
    pull(trait)))

scale <- 15 # scale for vector loadings

gift_pcoa_vectors %>% 
  rownames_to_column(var="genome") %>% 
  left_join(genome_metadata, by="genome") %>%
  mutate(relation = case_when(genome %in% taxonomy_urban_grey ~ "urban",
                              genome %in% taxonomy_rural_grey ~ "rural",
                              TRUE ~ "neutral")) %>%
 group_by(relation) %>%
  mutate(x_cen = mean(Axis.1, na.rm = TRUE)) %>%
  mutate(y_cen = mean(Axis.2, na.rm = TRUE)) %>%
  ungroup() %>% 
  mutate(phylum=ifelse(relation=="neutral","NA",phylum)) %>%
  arrange(phylum) %>%
  ggplot() +
      #genome positions
      scale_color_manual(values=custom_colors,na.value = "#EDEDED")+
      geom_point(aes(x=Axis.1,y=Axis.2, color=phylum, size=length), alpha=0.9, shape=16) +
      new_scale_color() +
      scale_color_manual(values = c("urban" = "#b592ad", "rural" = "#849c20")) + 
      geom_segment(data = . %>% filter(relation != "neutral"),
                   aes(x = x_cen, y = y_cen, xend = Axis.1, yend = Axis.2, color=relation), 
                   alpha = 0.5) +      
      new_scale_color() +
      #loading positions
      geom_segment(data=gift_pcoa_urb_grey, 
                   aes(x=0, y=0, xend=Axis.1 * scale, yend=Axis.2 * scale),
                    arrow = arrow(length = unit(0.3, "cm"), 
                    type = "open", 
                    angle = 25),
                    linewidth = 0.5, 
                    color = "black") +
     #Primary and secondary scale adjustments
     scale_x_continuous(name = paste0("PCoA1 (",round(gift_pcoa_rel_eigen[1]*100, digits = 2), " %)"),
                      sec.axis = sec_axis(~ . / scale, name = "Loadings on PCoA1")
            ) +
     scale_y_continuous(name = paste0("PCoA2 (",round(gift_pcoa_rel_eigen[2]*100, digits = 2), " %)"),
                      sec.axis = sec_axis(~ . / scale, name = "Loadings on PCoA2")
            ) +
    geom_label_repel(data = gift_pcoa_urb_grey,
                     aes(label = label, x = Axis.1 * scale, y = Axis.2 * scale),
                     segment.color = 'transparent') +
    theme_minimal() + 
    theme(legend.position = "none")

```


## Predictions by season

### Compositional predictions

```{r hmsc_season_enrichment_red, warning=FALSE, comments="", message=FALSE, fig.dim=c(8, 10)}

aut_enrichment_red <- post_estimates_red %>%
    filter(variable=="seasonautumn") %>%
    select(genome, trend)
win_enrichment_red <- post_estimates_red %>%
    filter(variable=="seasonwinter") %>%
    select(genome, trend)

taxonomy_spring_red <- post_estimates_red %>%
    filter(variable=="seasonwinter", 
           genome %in% pred_mags.red,
           trend =="Negative") %>%
    pull(genome)

taxonomy_win_red <- post_estimates_red %>%
    filter(variable=="seasonwinter", 
           genome %in% pred_mags.red,
           trend =="Positive") %>%
    pull(genome)

```


```{r season_prediction_red, warning=FALSE, comments="", message=FALSE}

gradient = c("spring-summer","autumn", "winter")
gradientlength = length(gradient)

pred_season_red <- constructGradient(m.red, 
                      focalVariable = "season", 
                      non.focalVariables = 1,
                      ngrid=gradientlength) %>%
            predict(m.red, Gradient = ., expected = TRUE) %>%
            do.call(rbind,.) %>%
            as.data.frame() %>%
            mutate(season=rep(gradient,1000)) %>%
            pivot_longer(!season,names_to = "genome", values_to = "value")

```


```{r season_plot_red, comment="", message=FALSE, warning=FALSE, fig.dim=c(8,8), fig.fullwidth=TRUE}


genome_metadata_clean <- genome_metadata %>%
  mutate(genus = if_else(!is.na(genus) & genus == "g__", paste0("unclassified ", family), genus))

# family level
pred_season_red %>%
  filter(genome %in% pred_mags.red) %>% #keep only predictive mags
  left_join(win_enrichment_red,by="genome") %>%
      filter(trend != "Neutral") %>%
      pivot_wider(names_from = season, values_from = value) %>%
      unnest(c(`winter`, `spring-summer`)) %>%
      mutate(diff_sw = `spring-summer` - `winter`) %>%
      select(genome, diff_sw) %>%
      left_join(genome_metadata_clean, by="genome") %>%
      mutate(genome= factor(genome, levels=genome_tree$tip.label)) %>%
      ggplot(., aes(y=forcats::fct_reorder(family,diff_sw), x=diff_sw, 
                    fill=phylum, color=phylum)) +
            geom_boxplot(outlier.shape = NA) +
            scale_color_manual(values=custom_colors) +
            scale_fill_manual(values=alpha(custom_colors,0.4)) +
            geom_vline(xintercept = 0) +
            annotate('text', x=-10, y=11, label = "Enriched in\nspring", color='black') + 
            annotate('text', x=10, y=11, label = "Enriched in\nwinter", color='black') +
            theme_classic()+
            labs(x = "Difference in log-abundance", y = "Genera", fill="Phylum", color="Phylum") +
            xlim(-16,16) 


#genus level
sw_red <-pred_season_red %>%
    filter(genome %in% pred_mags.red) %>% #keep only predictive mags
    left_join(win_enrichment_red,by="genome") %>%
      filter(trend != "Neutral") %>%
      pivot_wider(names_from = season, values_from = value) %>%
      unnest(c(`winter`, `spring-summer`)) %>%
      mutate(diff_sw = `spring-summer` - `winter`) %>%
      select(genome, diff_sw) %>%
      left_join(genome_metadata_clean, by="genome") %>%
      mutate(genome= factor(genome, levels=genome_tree$tip.label)) %>%
      ggplot(., aes(y=forcats::fct_reorder(genus,diff_sw), x=diff_sw, 
                    fill=factor(phylum, levels =c("p__Actinomycetota", "p__Bacillota","p__Bacillota_A", "p__Bacillota_B",
                                                   "p__Bacillota_C", "p__Bacteroidota", "p__Campylobacterota", "p__Cyanobacteriota",
                                                   "p__Pseudomonadota")), 
                    color=factor(phylum, levels =c("p__Actinomycetota", "p__Bacillota","p__Bacillota_A", "p__Bacillota_B",
                                                   "p__Bacillota_C", "p__Bacteroidota", "p__Campylobacterota", "p__Cyanobacteriota",
                                                   "p__Pseudomonadota")))) +
            geom_boxplot(outlier.shape = NA) +
            scale_color_manual(values=custom_colors, drop=FALSE)+
            scale_fill_manual(values=alpha(custom_colors,0.4), drop=FALSE)+
            annotate('text', x=-10, y=26, label = "Enriched in\nspring", color='black') + 
            annotate('text', x=10, y=26, label = "Enriched in\nwinter", color='black') +
            geom_vline(xintercept = 0)+
            theme_classic()+
            labs(x = "Difference in log-abundance", y = "Genera", fill="Phylum", color="Phylum") +
            xlim(-16,16) + 
            theme(plot.margin = margin(l = 35),
                  axis.title.x=element_blank(),
                  legend.position="none") 



```


```{r hmsc_season_enrichment_grey, warning=FALSE, comments="", message=FALSE}

aut_enrichment_grey <- post_estimates_grey %>%
    filter(variable=="seasonautumn") %>%
    select(genome, trend)
win_enrichment_grey <- post_estimates_grey %>%
    filter(variable=="seasonwinter") %>%
    select(genome, trend)

taxonomy_spring_grey <- post_estimates_grey %>%
    filter(variable=="seasonwinter", 
           genome %in% pred_mags.grey,
           trend =="Negative") %>%
    pull(genome)
taxonomy_win_grey <- post_estimates_grey %>%
    filter(variable=="seasonwinter", 
           genome %in% pred_mags.grey,
           trend =="Positive") %>%
    pull(genome)

# m$covNames

```


```{r season_prediction_grey, warning=FALSE, comments="", message=FALSE}

pred_season_grey <- constructGradient(m.grey, 
                      focalVariable = "season", 
                      non.focalVariables = 1,
                      ngrid=gradientlength) %>%
            predict(m.grey, Gradient = ., expected = TRUE) %>%
            do.call(rbind,.) %>%
            as.data.frame() %>%
            mutate(season=rep(gradient,1000)) %>%
            pivot_longer(!season,names_to = "genome", values_to = "value")
```


```{r season_plot_grey, comment="", message=FALSE, warning=FALSE, fig.dim=c(10,8), fig.fullwidth=TRUE}

#family level
pred_season_grey %>%
  filter(genome %in% pred_mags.grey) %>% #keep only predictive mags
  left_join(win_enrichment_grey,by="genome") %>%
      filter(trend != "Neutral") %>%
      pivot_wider(names_from = season, values_from = value) %>%
      unnest(c(`winter`, `spring-summer`)) %>%
      mutate(diff_sw = `spring-summer` - `winter`) %>%
      select(genome, diff_sw) %>%
      left_join(genome_metadata_clean, by="genome") %>%
      mutate(genome= factor(genome, levels=genome_tree$tip.label)) %>%
      ggplot(., aes(y=forcats::fct_reorder(family,diff_sw), x=diff_sw, 
                    fill=phylum, color=phylum)) +
            geom_boxplot(outlier.shape = NA) +
            scale_color_manual(values=custom_colors) +
            scale_fill_manual(values=alpha(custom_colors,0.4)) +
            geom_vline(xintercept = 0) +
            annotate('text', x=-10, y=8, label = "Enriched in\nspring", color='black') + 
            annotate('text', x=10, y=8, label = "Enriched in\nwinter", color='black') +
            theme_classic()+
            labs(x = "Difference in log-abundance", y = "Genera", fill="Phylum", color="Phylum") +
            xlim(-16,16) 

# genus level
sw_grey <- pred_season_grey %>%
  filter(genome %in% pred_mags.grey) %>% #keep only predictive mags
  left_join(win_enrichment_grey,by="genome") %>%
      filter(trend != "Neutral") %>%
      pivot_wider(names_from = season, values_from = value) %>%
      unnest(c(`winter`, `spring-summer`)) %>%
      mutate(diff_sw = `spring-summer` - `winter`) %>%
      select(genome, diff_sw) %>%
      left_join(genome_metadata_clean, by="genome") %>%
      mutate(genome= factor(genome, levels=genome_tree$tip.label)) %>%
      ggplot(., aes(y=forcats::fct_reorder(genus,diff_sw), x=diff_sw, 
                    fill=factor(phylum, levels =c("p__Actinomycetota", "p__Bacillota","p__Bacillota_A", "p__Bacillota_B",
                                                   "p__Bacillota_C", "p__Bacteroidota", "p__Campylobacterota", "p__Cyanobacteriota",
                                                   "p__Pseudomonadota")), 
                    color=factor(phylum, levels =c("p__Actinomycetota", "p__Bacillota","p__Bacillota_A", "p__Bacillota_B",
                                                   "p__Bacillota_C", "p__Bacteroidota", "p__Campylobacterota", "p__Cyanobacteriota",
                                                   "p__Pseudomonadota")))) +
            geom_boxplot(outlier.shape = NA) +
            scale_color_manual(values=custom_colors) +
            scale_fill_manual(values=alpha(custom_colors,0.4)) +
            geom_vline(xintercept = 0) +
            annotate('text', x=-10, y=17, label = "Enriched in\nspring", color='black') + 
            annotate('text', x=10, y=17, label = "Enriched in\nwinter", color='black') +
            theme_classic()+
            labs(x = "Difference in log-abundance", y = "Genera", fill="Phylum", color="Phylum") +
            xlim(-16,16) 


     

```

```{r seas_plot_all, fig.dim=c(10,10)}

ggarrange(sw_red, sw_grey, legend="right", common.legend = TRUE,
          #labels = c("A", "B", "C"),
          ncol = 1, nrow = 2, align="v")

```


### Functional predictions


```{r function_seas_pred_red, results="hide", eval=TRUE, message=FALSE, warning=FALSE, cache=TRUE}

community_func_seas_red <- pred_season_red  %>%
  filter(genome %in% pred_mags.red) %>% #keep only predictive mags
  group_by(season, genome) %>%
  mutate(row_id = row_number()) %>%
  pivot_wider(names_from = genome, values_from = value) %>%
  ungroup() %>%
  group_split(row_id) %>%
  as.list() %>%
  lapply(., FUN = function(x){x %>%
    select(-row_id) %>%
    column_to_rownames(var = "season") %>%
    as.data.frame() %>%
    exp() %>%
    t() %>%
    tss() %>%
    to.community(GIFTs_functions,.,GIFT_db2) %>%
    as.data.frame() %>%
    rownames_to_column(var="season")
   })

function_pred_seas_red <- map_dfr(community_func_seas_red, function(df) {
  spring_values <- df %>% filter(season == "spring-summer") %>% select(-season)
  winter_values <- df %>% filter(season == "winter") %>% select(-season)
  spring_values - winter_values
}) %>%
  mutate(iteration=c(1:1000)) %>% 
  pivot_longer(!iteration,names_to="trait",values_to="value") %>%
      group_by(trait) %>%
      summarise(mean=mean(value),
        p1 = quantile(value, probs = 0.1),
        p9 = quantile(value, probs = 0.9),
        positive_support = sum(value > 0)/1000,
        negative_support = sum(value < 0)/1000) %>%
      arrange(-positive_support)


```

```{r function_seas_plot_red, message=FALSE, warning=FALSE, fig.dim=c(10, 8)}

# Winter associated
function_pred_seas_red %>%
  filter(mean >0, positive_support>=0.9) %>%
  arrange(-mean) %>%
  select(-negative_support) %>%
  rename(support=positive_support) %>%
  left_join(GIFT_db2 %>% select(Code_function, Function),by=join_by(trait==Code_function)) %>%
  relocate(Function, .after=trait) %>%
  unique() %>%
  paged_table()

# Spring associated
function_pred_seas_red %>%
  filter(mean <0, negative_support>=0.9) %>%
  arrange(mean) %>%
  select(-positive_support) %>%
  rename(support=negative_support) %>%
  left_join(GIFT_db2 %>% select(Code_function, Function),by=join_by(trait==Code_function)) %>%
  relocate(Function, .after=trait) %>%
  unique() %>%
  paged_table()


all_functions_red <- function_pred_seas_red %>%
  left_join(GIFT_db2,by=join_by(trait==Code_function)) %>%
  mutate(trait=factor(trait)) %>%
  mutate(function_legend=str_c(trait," - ",Function)) %>%
  select(trait,mean,p1,p9,function_legend) %>% 
  unique()

red_fun_seas_plot <- all_functions_red %>%
  ggplot(aes(x=mean, y=function_legend, xmin=p1, xmax=p9, color=function_legend)) +
      geom_point() +
      geom_errorbar() +
      xlim(c(-3,3)) +
      geom_vline(xintercept=0) +
      scale_color_manual(values = gift_colors$Color) +
      theme_minimal() +
      labs(x="Difference in log-abundance",y="Functional trait", title="S. vulgaris") +
      annotate('text', x=-2, y=12, label = "Enriched in\nspring", color='black') +
      annotate('text', x=2, y=12, label = "Enriched in\nwinter", color='black') +
      #guides(col = guide_legend(ncol = 1))
      theme(legend.position = "none",
            plot.title=element_text( hjust=0.5, vjust=0.6, face='bold', size=12))

```


```{r function_seas_pred_grey, results="hide", eval=TRUE, message=FALSE, warning=FALSE, cache=TRUE}

community_func_seas_grey <- pred_season_grey  %>%
  filter(genome %in% pred_mags.grey) %>% #keep only predictive mags
  group_by(season, genome) %>%
  mutate(row_id = row_number()) %>%
  pivot_wider(names_from = genome, values_from = value) %>%
  ungroup() %>%
  group_split(row_id) %>%
  as.list() %>%
  lapply(., FUN = function(x){x %>%
    select(-row_id) %>%
    column_to_rownames(var = "season") %>%
    as.data.frame() %>%
    exp() %>%
    t() %>%
    tss() %>%
    to.community(GIFTs_functions,.,GIFT_db2) %>%
    as.data.frame() %>%
    rownames_to_column(var="season")
   })

function_pred_seas_grey <- map_dfr(community_func_seas_grey, function(df) {
  spring_values <- df %>% filter(season == "spring-summer") %>% select(-season)
  winter_values <- df %>% filter(season == "winter") %>% select(-season)
  spring_values - winter_values
}) %>%
  mutate(iteration=c(1:1000)) %>% 
  pivot_longer(!iteration,names_to="trait",values_to="value") %>%
      group_by(trait) %>%
      summarise(mean=mean(value),
        p1 = quantile(value, probs = 0.1),
        p9 = quantile(value, probs = 0.9),
        positive_support = sum(value > 0)/1000,
        negative_support = sum(value < 0)/1000) %>%
      arrange(-positive_support)


```

```{r function_seas_plot_grey, message=FALSE, warning=FALSE, fig.dim=c(10, 8)}

# Winter associated
function_pred_seas_grey %>%
  filter(mean >0, positive_support>=0.9) %>%
  arrange(-mean) %>%
  select(-negative_support) %>%
  rename(support=positive_support) %>%
  left_join(GIFT_db2 %>% select(Code_function, Function),by=join_by(trait==Code_function)) %>%
  relocate(Function, .after=trait) %>%
  unique() %>%
  paged_table()

# Spring associated
function_pred_seas_grey %>%
  filter(mean <0, negative_support>=0.9) %>%
  arrange(mean) %>%
  select(-positive_support) %>%
  rename(support=negative_support) %>%
  left_join(GIFT_db2 %>% select(Code_function, Function),by=join_by(trait==Code_function)) %>%
  relocate(Function, .after=trait) %>%
  unique() %>%
  paged_table()


all_functions_grey <- function_pred_seas_grey %>%
  left_join(GIFT_db2,by=join_by(trait==Code_function)) %>%
  mutate(trait=factor(trait)) %>%
  mutate(function_legend=str_c(trait," - ",Function)) %>%
  select(trait,mean,p1,p9,function_legend) %>% 
  unique()

grey_fun_seas_plot <- all_functions_grey %>%
  ggplot(aes(x=mean, y=function_legend, xmin=p1, xmax=p9, color=function_legend)) +
      geom_point() +
      geom_errorbar() +
      xlim(c(-3,3)) +
      geom_vline(xintercept=0) +
      scale_color_manual(values = gift_colors$Color) +
      theme_minimal() +
      labs(x="Difference in log-abundance",y="Functional trait", title="S. carolinensis") +
      annotate('text', x=-2, y=12, label = "Enriched in\nspring", color='black') +
      annotate('text', x=2, y=12, label = "Enriched in\nwinter", color='black') +
      #guides(col = guide_legend(ncol = 1))
      theme(legend.position = "none",
            axis.title.y = element_blank(),
            axis.text.y = element_blank(),
            plot.title=element_text( hjust=0.5, vjust=0.6, face='bold', size=12))



```


```{r fun_seas_plot, message=FALSE, warning=FALSE, fig.dim=c(10, 8)}

ggarrange(red_fun_seas_plot, grey_fun_seas_plot, 
          ncol=2, nrow=1, align='h')

```



```{r element_seas_pred_red, results="hide", eval=TRUE, message=FALSE, warning=FALSE, cache=TRUE}

community_elem_seas_red <- pred_season_red  %>%
  filter(genome %in% pred_mags.red) %>% #keep only predictive mags
  group_by(season, genome) %>%
  mutate(row_id = row_number()) %>%
  pivot_wider(names_from = genome, values_from = value) %>%
  ungroup() %>%
  group_split(row_id) %>%
  as.list() %>%
  lapply(., FUN = function(x){x %>%
    select(-row_id) %>%
    column_to_rownames(var = "season") %>%
    as.data.frame() %>%
    exp() %>%
    t() %>%
    tss() %>%
    to.community(GIFTs_elements,.,GIFT_db2) %>%
    as.data.frame() %>%
    rownames_to_column(var="season")
   })

element_pred_seas_red <- map_dfr(community_elem_seas_red, function(df) {
  spring_values <- df %>% filter(season == "spring-summer") %>% select(-season)
  winter_values <- df %>% filter(season == "winter") %>% select(-season)
  spring_values - winter_values
}) %>%
  mutate(iteration=c(1:1000)) %>% 
  pivot_longer(!iteration,names_to="trait",values_to="value") %>%
      group_by(trait) %>%
      summarise(mean=mean(value),
        p1 = quantile(value, probs = 0.1),
        p9 = quantile(value, probs = 0.9),
        positive_support = sum(value > 0)/1000,
        negative_support = sum(value < 0)/1000) %>%
      arrange(-positive_support)



```

```{r element_seas_plot_red, message=FALSE, warning=FALSE, fig.dim=c(10, 8)}

# Winter associated
element_pred_seas_red %>%
  filter(mean >0, positive_support>=0.9) %>%
  arrange(-mean) %>%
  select(-negative_support) %>%
  rename(support=positive_support) %>%
  left_join(GIFT_db2 %>% select(Code_element, Element, Function),by=join_by(trait==Code_element)) %>%
  relocate(Element, Function, .after=trait) %>%
  unique() %>%
  paged_table()

# Spring associated
element_pred_seas_red %>%
  filter(mean <0, negative_support>=0.9) %>%
  arrange(mean) %>%
  select(-positive_support) %>%
  rename(support=negative_support) %>%
  left_join(GIFT_db2 %>% select(Code_element, Element, Function),by=join_by(trait==Code_element)) %>%
  relocate(Element, Function, .after=trait) %>%
  unique() %>%
  paged_table()


element_winter_red <- element_pred_seas_red%>%
  filter(mean >0) %>%
  arrange(mean) %>%
  filter(positive_support>=0.9) %>%
  select(-negative_support) %>%
  rename(support=positive_support)
element_spring_red <- element_pred_seas_red %>%
  filter(mean <0) %>%
  arrange(mean) %>%
  filter(negative_support>=0.9) %>%
  select(-positive_support) %>%
  rename(support=negative_support)

all_elements_red <- bind_rows(element_spring_red,element_winter_red) %>%
  left_join(GIFT_db2,by=join_by(trait==Code_element)) %>%
  mutate(trait=factor(trait,levels=c(rev(element_spring_red$trait),rev(element_winter_red$trait)))) %>%
  mutate(Code_function=factor(Code_function)) %>%
  mutate(element_legend=str_c(trait," - ",Element)) %>%
  mutate(function_legend=str_c(Code_function," - ",Function)) %>%
  select(trait,mean,p1,p9,element_legend,function_legend) %>% 
  unique()

red_elem_seas_plot <- all_elements_red %>%
  ggplot(aes(x=mean, y=fct_reorder(element_legend, mean), xmin=p1, xmax=p9, color=function_legend)) +
      geom_point() +
      geom_errorbar() +
      xlim(c(-3.5,3.5)) +
      geom_vline(xintercept=0) +
      scale_color_manual(values = gift_colors$Color) +
      theme_minimal() +
      labs(x="Difference in log-abundance",y="Functional trait", title = "S. vulgaris", color="Function") +
      annotate('text', x=-2, y=12, label = "Enriched in\nspring", color='black') +  #NB reversed
      annotate('text', x=2, y=12, label = "Enriched in\nwinter", color='black') +
      #guides(col = guide_legend(ncol = 1))
      theme(legend.position = "right",
            plot.title=element_text( hjust=0.45, vjust=0.6, face='bold', size=12))



```


```{r element_seas_pred_grey, results="hide", eval=TRUE, message=FALSE, warning=FALSE, cache=TRUE}

community_elem_seas_grey <- pred_season_grey  %>%
  filter(genome %in% pred_mags.grey) %>% #keep only predictive mags
  group_by(season, genome) %>%
  mutate(row_id = row_number()) %>%
  pivot_wider(names_from = genome, values_from = value) %>%
  ungroup() %>%
  group_split(row_id) %>%
  as.list() %>%
  lapply(., FUN = function(x){x %>%
    select(-row_id) %>%
    column_to_rownames(var = "season") %>%
    as.data.frame() %>%
    exp() %>%
    t() %>%
    tss() %>%
    to.community(GIFTs_elements,.,GIFT_db2) %>%
    as.data.frame() %>%
    rownames_to_column(var="season")
   })

element_pred_seas_grey <- map_dfr(community_elem_seas_grey, function(df) {
  spring_values <- df %>% filter(season == "spring-summer") %>% select(-season)
  winter_values <- df %>% filter(season == "winter") %>% select(-season)
  spring_values - winter_values
}) %>%
  mutate(iteration=c(1:1000)) %>% 
  pivot_longer(!iteration,names_to="trait",values_to="value") %>%
      group_by(trait) %>%
      summarise(mean=mean(value),
        p1 = quantile(value, probs = 0.1),
        p9 = quantile(value, probs = 0.9),
        positive_support = sum(value > 0)/1000,
        negative_support = sum(value < 0)/1000) %>%
      arrange(-positive_support)



```

```{r element_seas_plot_grey, message=FALSE, warning=FALSE, fig.dim=c(10, 8)}

# Winter associated
element_pred_seas_grey %>%
  filter(mean >0, positive_support>=0.9) %>%
  arrange(-mean) %>%
  select(-negative_support) %>%
  rename(support=positive_support) %>%
  left_join(GIFT_db2 %>% select(Code_element, Element, Function),by=join_by(trait==Code_element)) %>%
  relocate(Element, Function, .after=trait) %>%
  unique() %>%
  paged_table()

# Spring associated
element_pred_seas_grey %>%
  filter(mean <0, negative_support>=0.9) %>%
  arrange(mean) %>%
  select(-positive_support) %>%
  rename(support=negative_support) %>%
  left_join(GIFT_db2 %>% select(Code_element, Element, Function),by=join_by(trait==Code_element)) %>%
  relocate(Element, Function, .after=trait) %>%
  unique() %>%
  paged_table()


element_win_grey <- element_pred_seas_grey%>%
  filter(mean >0) %>%
  arrange(mean) %>%
  filter(positive_support>=0.9) %>%
  select(-negative_support) %>%
  rename(support=positive_support)
element_spring_grey <- element_pred_seas_grey %>%
  filter(mean <0) %>%
  arrange(mean) %>%
  filter(negative_support>=0.9) %>%
  select(-positive_support) %>%
  rename(support=negative_support)

all_elements_grey <- bind_rows(element_spring_grey,element_win_grey ) %>%
  left_join(GIFT_db2,by=join_by(trait==Code_element)) %>%
  mutate(trait=factor(trait,levels=c(rev(element_spring_grey$trait),rev(element_win_grey$trait)))) %>%
  mutate(Code_function=factor(Code_function)) %>%
  mutate(element_legend=str_c(trait," - ",Element)) %>%
  mutate(function_legend=str_c(Code_function," - ",Function)) %>%
  select(trait,mean,p1,p9,element_legend,function_legend) %>% 
  unique()

grey_elem_seas_plot <- all_elements_grey %>%
  ggplot(aes(x=mean, y=fct_reorder(element_legend, mean), xmin=p1, xmax=p9, color=function_legend)) +
      geom_point() +
      geom_errorbar() +
      xlim(c(-3.5,3.5)) +
      geom_vline(xintercept=0) +
      scale_color_manual(values = gift_colors$Color) +
      theme_minimal() +
      labs(x="Difference in log-abundance",y="Functional trait", title ="S. carolinensis", color="Function") +
      annotate('text', x=-2, y=20, label = "Enriched in\nspring", color='black') + #NB reversed
      annotate('text', x=2, y=20, label = "Enriched in\nwinter", color='black') +
      #guides(col = guide_legend(ncol = 1))
      theme(legend.position = "right",
            axis.title.y = element_blank(),
            plot.title=element_text( hjust=0.5, vjust=0.6, face='bold', size=12))


```


```{r elem_seas_plot, message=FALSE, warning=FALSE, fig.dim=c(10, 10)}

ggarrange(red_elem_seas_plot, grey_elem_seas_plot, legend="bottom", common.legend=TRUE,
          ncol=2, nrow=1)

```

### Sources of functional variation

```{r ord_seas_red, warning=FALSE, message=FALSE}

gift_pcoa_seas_red <- cov(genome_gifts, scale(gift_pcoa_vectors)) %*% diag((gift_pcoa_eigenvalues/(nrow(genome_gifts)-1))^(-0.5)) %>%
  as.data.frame() %>% 
  rename(Axis.1=1,Axis.2=2) %>% 
  rownames_to_column(var="label") %>% 
  #get function summary vectors
  mutate(func=substr(label,1,5)) %>% 
  group_by(func) %>% 
  summarise(Axis.1=mean(Axis.1),
            Axis.2=mean(Axis.2)) %>% 
  rename(label=func) %>% 
  filter(label %in% (element_spring_red %>% 
    pull(trait)))


scale <- 15 # scale for vector loadings

gift_pcoa_vectors %>% 
  rownames_to_column(var="genome") %>% 
  left_join(genome_metadata, by="genome") %>%
  mutate(relation = case_when(genome %in% taxonomy_spring_red ~ "spring",
                              genome %in% taxonomy_win_red ~ "winter",
                              TRUE ~ "neutral")) %>%
 group_by(relation) %>%
  mutate(x_cen = mean(Axis.1, na.rm = TRUE)) %>%
  mutate(y_cen = mean(Axis.2, na.rm = TRUE)) %>%
  ungroup() %>% 
  mutate(phylum=ifelse(relation=="neutral","NA",phylum)) %>% 
  arrange(phylum)%>%
  ggplot() +
      #genome positions
      scale_color_manual(values=custom_colors,na.value = "#EDEDED")+
      geom_point(aes(x=Axis.1,y=Axis.2, color=phylum, size=length), alpha=0.9, shape=16) +
      new_scale_color() +
      scale_color_manual(values = c("winter" = "#b592ad", "spring" = "#849c20")) + 
      geom_segment(data = . %>% filter(relation != "neutral"),
                   aes(x = x_cen, y = y_cen, xend = Axis.1, yend = Axis.2, color=relation), 
                   alpha = 0.5) +      
      new_scale_color() +
      #loading positions
      geom_segment(data=gift_pcoa_seas_red, 
                   aes(x=0, y=0, xend=Axis.1 * scale, yend=Axis.2 * scale),
                    arrow = arrow(length = unit(0.3, "cm"), 
                    type = "open", 
                    angle = 25),
                    linewidth = 0.5, 
                    color = "black") +
     #Primary and secondary scale adjustments
     scale_x_continuous(name = paste0("PCoA1 (",round(gift_pcoa_rel_eigen[1]*100, digits = 2), " %)"),
                      sec.axis = sec_axis(~ . / scale, name = "Loadings on PCoA1")
            ) +
     scale_y_continuous(name = paste0("PCoA2 (",round(gift_pcoa_rel_eigen[2]*100, digits = 2), " %)"),
                      sec.axis = sec_axis(~ . / scale, name = "Loadings on PCoA2")
            ) +
    geom_label_repel(data = gift_pcoa_seas_red,
                     aes(label = label, x = Axis.1 * scale, y = Axis.2 * scale),
                     segment.color = 'transparent') +
    theme_minimal() + 
    theme(legend.position = "none")

```

```{r ord_seas_grey, warning=FALSE, message=FALSE}

gift_pcoa_seas_grey <- cov(genome_gifts, scale(gift_pcoa_vectors)) %*% diag((gift_pcoa_eigenvalues/(nrow(genome_gifts)-1))^(-0.5)) %>%
  as.data.frame() %>% 
  rename(Axis.1=1,Axis.2=2) %>% 
  rownames_to_column(var="label") %>% 
  #get function summary vectors
  mutate(func=substr(label,1,5)) %>% 
  group_by(func) %>% 
  summarise(Axis.1=mean(Axis.1),
            Axis.2=mean(Axis.2)) %>% 
  rename(label=func) %>% 
  filter(label %in% (element_spring_grey %>% 
    pull(trait)))

scale <- 15 # scale for vector loadings

gift_pcoa_vectors %>% 
  rownames_to_column(var="genome") %>% 
  left_join(genome_metadata, by="genome") %>%
  mutate(relation = case_when(genome %in% taxonomy_spring_grey ~ "spring",
                              genome %in% taxonomy_win_grey ~ "winter",
                              TRUE ~ "neutral")) %>%
 group_by(relation) %>%
  mutate(x_cen = mean(Axis.1, na.rm = TRUE)) %>%
  mutate(y_cen = mean(Axis.2, na.rm = TRUE)) %>%
  ungroup() %>% 
  mutate(phylum=ifelse(relation=="neutral","NA",phylum)) %>%
  arrange(phylum) %>%
  ggplot() +
      #genome positions
      scale_color_manual(values=custom_colors,na.value = "#EDEDED")+
      geom_point(aes(x=Axis.1,y=Axis.2, color=phylum, size=length), alpha=0.9, shape=16) +
      new_scale_color() +
      scale_color_manual(values = c("winter" = "#b592ad", "spring" = "#849c20")) + 
      geom_segment(data = . %>% filter(relation != "neutral"),
                   aes(x = x_cen, y = y_cen, xend = Axis.1, yend = Axis.2, color=relation), 
                   alpha = 0.5) +      
      new_scale_color() +
      #loading positions
      geom_segment(data=gift_pcoa_seas_grey, 
                   aes(x=0, y=0, xend=Axis.1 * scale, yend=Axis.2 * scale),
                    arrow = arrow(length = unit(0.3, "cm"), 
                    type = "open", 
                    angle = 25),
                    linewidth = 0.5, 
                    color = "black") +
     #Primary and secondary scale adjustments
     scale_x_continuous(name = paste0("PCoA1 (",round(gift_pcoa_rel_eigen[1]*100, digits = 2), " %)"),
                      sec.axis = sec_axis(~ . / scale, name = "Loadings on PCoA1")
            ) +
     scale_y_continuous(name = paste0("PCoA2 (",round(gift_pcoa_rel_eigen[2]*100, digits = 2), " %)"),
                      sec.axis = sec_axis(~ . / scale, name = "Loadings on PCoA2")
            ) +
    geom_label_repel(data = gift_pcoa_seas_grey,
                     aes(label = label, x = Axis.1 * scale, y = Axis.2 * scale),
                     segment.color = 'transparent') +
    theme_minimal() + 
    theme(legend.position = "none")

```

## Predictions by sex

### Compositional predictions

```{r hmsc_sex_enrichment_red, warning=FALSE, comments="", message=FALSE, fig.dim=c(8, 10)}

male_enrichment_red <- post_estimates_red %>%
    filter(variable=="sexM") %>%
    select(genome, trend)


taxonomy_female_red <- post_estimates_red %>%
    filter(variable=="sexM", 
           genome %in% pred_mags.red,
           trend =="Negative") %>%
    pull(genome)

taxonomy_male_red <- post_estimates_red %>%
    filter(variable=="sexM", 
           genome %in% pred_mags.red,
           trend =="Positive") %>%
    pull(genome)

```


```{r sex_prediction_red, warning=FALSE, comments="", message=FALSE}

gradient = c("female","male")
gradientlength = length(gradient)

pred_sex_red <- constructGradient(m.red, 
                      focalVariable = "sex", 
                      non.focalVariables = 1,
                      ngrid=gradientlength) %>%
            predict(m.red, Gradient = ., expected = TRUE) %>%
            do.call(rbind,.) %>%
            as.data.frame() %>%
            mutate(sex=rep(gradient,1000)) %>%
            pivot_longer(!sex,names_to = "genome", values_to = "value")

```


```{r sex_plot_red, comment="", message=FALSE, warning=FALSE, fig.dim=c(8,8), fig.fullwidth=TRUE}


fm_red <-pred_sex_red %>%
    filter(genome %in% pred_mags.red) %>% #keep only predictive mags
    left_join(male_enrichment_red,by="genome") %>%
      filter(trend != "Neutral") %>%
      pivot_wider(names_from = sex, values_from = value) %>%
      unnest(c(`male`, `female`)) %>%
      mutate(diff_fm = `female` - `male`) %>%
      select(genome, diff_fm) %>%
      left_join(genome_metadata_clean, by="genome") %>%
      mutate(genome= factor(genome, levels=genome_tree$tip.label)) %>%
      ggplot(., aes(y=forcats::fct_reorder(genus,diff_fm), x=diff_fm, fill=phylum, color=phylum)) +
            geom_boxplot(outlier.shape = NA) +
            scale_color_manual(values=custom_colors)+
            scale_fill_manual(values=alpha(custom_colors,0.4))+
            annotate('text', x=-10, y=15, label = "Enriched in\nfemales", color='black') + #NB: reversed because it makes no sense, needs to be checked
            annotate('text', x=10, y=15, label = "Enriched in\nmales", color='black') +
            geom_vline(xintercept = 0)+
            theme_classic()+
            labs(x = "Difference in log-abundance", y = "Genera") +
            xlim(-15,15) + 
            theme(plot.margin = margin(l = 35),
                  axis.title.x=element_blank()) 


```


```{r hmsc_sex_enrichment_grey, warning=FALSE, comments="", message=FALSE, fig.dim=c(8, 10)}

male_enrichment_grey <- post_estimates_grey %>%
    filter(variable=="sexM") %>%
    select(genome, trend)


taxonomy_female_grey <- post_estimates_grey %>%
    filter(variable=="sexM", 
           genome %in% pred_mags.grey,
           trend =="Negative") %>%
    pull(genome)

taxonomy_male_grey <- post_estimates_grey %>%
    filter(variable=="sexM", 
           genome %in% pred_mags.grey,
           trend =="Positive") %>%
    pull(genome)

```


```{r sex_prediction_grey, warning=FALSE, comments="", message=FALSE}

gradient = c("female","male")
gradientlength = length(gradient)

pred_sex_grey <- constructGradient(m.grey, 
                      focalVariable = "sex", 
                      non.focalVariables = 1,
                      ngrid=gradientlength) %>%
            predict(m.grey, Gradient = ., expected = TRUE) %>%
            do.call(rbind,.) %>%
            as.data.frame() %>%
            mutate(sex=rep(gradient,1000)) %>%
            pivot_longer(!sex,names_to = "genome", values_to = "value")

```


```{r sex_plot_grey, comment="", message=FALSE, warning=FALSE, fig.dim=c(8,8), fig.fullwidth=TRUE}


fm_grey <-pred_sex_grey %>%
    filter(genome %in% pred_mags.grey) %>% #keep only predictive mags
    left_join(male_enrichment_grey,by="genome") %>%
      filter(trend != "Neutral") %>%
      pivot_wider(names_from = sex, values_from = value) %>%
      unnest(c(`male`, `female`)) %>%
      mutate(diff_fm = `female` - `male`) %>%
      select(genome, diff_fm) %>%
      left_join(genome_metadata_clean, by="genome") %>%
      mutate(genome= factor(genome, levels=genome_tree$tip.label)) %>%
      ggplot(., aes(y=forcats::fct_reorder(genus,diff_fm), x=diff_fm, fill=phylum, color=phylum)) +
            geom_boxplot(outlier.shape = NA) +
            scale_color_manual(values=custom_colors)+
            scale_fill_manual(values=alpha(custom_colors,0.4))+
            annotate('text', x=-10, y=7, label = "Enriched in\nfemales", color='black') + #NB: reversed because it makes no sense, needs to be checked
            annotate('text', x=10, y=7, label = "Enriched in\nmales", color='black') +
            geom_vline(xintercept = 0)+
            theme_classic()+
            labs(x = "Difference in log-abundance", y = "Genera") +
            xlim(-15,15) + 
            theme(plot.margin = margin(l = 35),
                  axis.title.x=element_blank()) 


```

```{r sex_plot_all, fig.dim=c(10,10)}

ggarrange(fm_red, fm_grey, legend="right", common.legend = TRUE,
          #labels = c("A", "B", "C"),
          ncol = 1, nrow = 2, align="v")

```

### Functional predictions

```{r function_sex_pred_red, results="hide", eval=TRUE, message=FALSE, warning=FALSE, cache=TRUE}

community_func_sex_red <- pred_sex_red  %>%
  filter(genome %in% pred_mags.red) %>% #keep only predictive mags
  group_by(sex, genome) %>%
  mutate(row_id = row_number()) %>%
  pivot_wider(names_from = genome, values_from = value) %>%
  ungroup() %>%
  group_split(row_id) %>%
  as.list() %>%
  lapply(., FUN = function(x){x %>%
    select(-row_id) %>%
    column_to_rownames(var = "sex") %>%
    as.data.frame() %>%
    exp() %>%
    t() %>%
    tss() %>%
    to.community(GIFTs_functions,.,GIFT_db2) %>%
    as.data.frame() %>%
    rownames_to_column(var="sex")
   })

function_pred_sex_red <- map_dfr(community_func_sex_red, function(df) {
  f_values <- df %>% filter(sex == "female") %>% select(-sex)
  m_values <- df %>% filter(sex == "male") %>% select(-sex)
  f_values - m_values
}) %>%
  mutate(iteration=c(1:1000)) %>% 
  pivot_longer(!iteration,names_to="trait",values_to="value") %>%
      group_by(trait) %>%
      summarise(mean=mean(value),
        p1 = quantile(value, probs = 0.1),
        p9 = quantile(value, probs = 0.9),
        positive_support = sum(value > 0)/1000,
        negative_support = sum(value < 0)/1000) %>%
      arrange(-positive_support)


```

```{r function_sex_plot_red, message=FALSE, warning=FALSE, fig.dim=c(10, 8)}

# male associated
function_pred_sex_red %>%
  filter(mean >0, positive_support>=0.9) %>%
  arrange(-mean) %>%
  select(-negative_support) %>%
  rename(support=positive_support) %>%
  left_join(GIFT_db2 %>% select(Code_function, Function),by=join_by(trait==Code_function)) %>%
  relocate(Function, .after=trait) %>%
  unique() %>%
  paged_table()

# female associated
function_pred_sex_red %>%
  filter(mean <0, negative_support>=0.9) %>%
  arrange(mean) %>%
  select(-positive_support) %>%
  rename(support=negative_support) %>%
  left_join(GIFT_db2 %>% select(Code_function, Function),by=join_by(trait==Code_function)) %>%
  relocate(Function, .after=trait) %>%
  unique() %>%
  paged_table()


all_functions_red <- function_pred_sex_red %>%
  left_join(GIFT_db2,by=join_by(trait==Code_function)) %>%
  mutate(trait=factor(trait)) %>%
  mutate(function_legend=str_c(trait," - ",Function)) %>%
  select(trait,mean,p1,p9,function_legend) %>% 
  unique()

red_fun_sex_plot <- all_functions_red %>%
  ggplot(aes(x=mean, y=function_legend, xmin=p1, xmax=p9, color=function_legend)) +
      geom_point() +
      geom_errorbar() +
      xlim(c(-2.5,2.5)) +
      geom_vline(xintercept=0) +
      scale_color_manual(values = gift_colors$Color) +
      theme_minimal() +
      labs(x="Difference in log-abundance",y="Functional trait", title="S. vulgaris") +
      annotate('text', x=-1.5, y=10, label = "Enriched in\nfemales", color='black') +
      annotate('text', x=1.5, y=10, label = "Enriched in\nmales", color='black') +
      #guides(col = guide_legend(ncol = 1))
      theme(legend.position = "none",
            plot.title=element_text( hjust=0.5, vjust=0.6, face='bold', size=12))

```
```{r function_sex_pred_grey, results="hide", eval=TRUE, message=FALSE, warning=FALSE, cache=TRUE}

community_func_sex_grey <- pred_sex_grey  %>%
  filter(genome %in% pred_mags.red) %>% #keep only predictive mags
  group_by(sex, genome) %>%
  mutate(row_id = row_number()) %>%
  pivot_wider(names_from = genome, values_from = value) %>%
  ungroup() %>%
  group_split(row_id) %>%
  as.list() %>%
  lapply(., FUN = function(x){x %>%
    select(-row_id) %>%
    column_to_rownames(var = "sex") %>%
    as.data.frame() %>%
    exp() %>%
    t() %>%
    tss() %>%
    to.community(GIFTs_functions,.,GIFT_db2) %>%
    as.data.frame() %>%
    rownames_to_column(var="sex")
   })

function_pred_sex_grey <- map_dfr(community_func_sex_grey, function(df) {
  f_values <- df %>% filter(sex == "female") %>% select(-sex)
  m_values <- df %>% filter(sex == "male") %>% select(-sex)
  f_values - m_values
}) %>%
  mutate(iteration=c(1:1000)) %>% 
  pivot_longer(!iteration,names_to="trait",values_to="value") %>%
      group_by(trait) %>%
      summarise(mean=mean(value),
        p1 = quantile(value, probs = 0.1),
        p9 = quantile(value, probs = 0.9),
        positive_support = sum(value > 0)/1000,
        negative_support = sum(value < 0)/1000) %>%
      arrange(-positive_support)


```


```{r function_sex_plot_grey, message=FALSE, warning=FALSE, fig.dim=c(10, 8)}

# male associated
function_pred_sex_grey %>%
  filter(mean >0, positive_support>=0.9) %>%
  arrange(-mean) %>%
  select(-negative_support) %>%
  rename(support=positive_support) %>%
  left_join(GIFT_db2 %>% select(Code_function, Function),by=join_by(trait==Code_function)) %>%
  relocate(Function, .after=trait) %>%
  unique() %>%
  paged_table()

# female associated
function_pred_sex_grey %>%
  filter(mean <0, negative_support>=0.9) %>%
  arrange(mean) %>%
  select(-positive_support) %>%
  rename(support=negative_support) %>%
  left_join(GIFT_db2 %>% select(Code_function, Function),by=join_by(trait==Code_function)) %>%
  relocate(Function, .after=trait) %>%
  unique() %>%
  paged_table()


all_functions_grey <- function_pred_sex_grey %>%
  left_join(GIFT_db2,by=join_by(trait==Code_function)) %>%
  mutate(trait=factor(trait)) %>%
  mutate(function_legend=str_c(trait," - ",Function)) %>%
  select(trait,mean,p1,p9,function_legend) %>% 
  unique()

grey_fun_sex_plot <- all_functions_grey %>%
  ggplot(aes(x=mean, y=function_legend, xmin=p1, xmax=p9, color=function_legend)) +
      geom_point() +
      geom_errorbar() +
      xlim(c(-2.5,2.5)) +
      geom_vline(xintercept=0) +
      scale_color_manual(values = gift_colors$Color) +
      theme_minimal() +
      labs(x="Difference in log-abundance",y="Functional trait", title="S. vulgaris") +
      annotate('text', x=-1.5, y=10, label = "Enriched in\nfemales", color='black') +
      annotate('text', x=1.5, y=10, label = "Enriched in\nmales", color='black') +
      #guides(col = guide_legend(ncol = 1))
      theme(legend.position = "none",
            plot.title=element_text( hjust=0.5, vjust=0.6, face='bold', size=12))

```

```{r fun_sex_plot_all, fig.dim=c(10,10)}

ggarrange(red_fun_sex_plot, grey_fun_sex_plot, legend="right", common.legend = TRUE,
          #labels = c("A", "B", "C"),
          ncol = 1, nrow = 2, align="v")

```

```{r element_sex_pred_red, results="hide", eval=TRUE, message=FALSE, warning=FALSE, cache=TRUE}

community_elem_sex_red <- pred_sex_red  %>%
  filter(genome %in% pred_mags.red) %>% #keep only predictive mags
  group_by(sex, genome) %>%
  mutate(row_id = row_number()) %>%
  pivot_wider(names_from = genome, values_from = value) %>%
  ungroup() %>%
  group_split(row_id) %>%
  as.list() %>%
  lapply(., FUN = function(x){x %>%
    select(-row_id) %>%
    column_to_rownames(var = "sex") %>%
    as.data.frame() %>%
    exp() %>%
    t() %>%
    tss() %>%
    to.community(GIFTs_elements,.,GIFT_db2) %>%
    as.data.frame() %>%
    rownames_to_column(var="sex")
   })

element_pred_sex_red <- map_dfr(community_elem_sex_red, function(df) {
  f_values <- df %>% filter(sex == "female") %>% select(-sex)
  m_values <- df %>% filter(sex == "male") %>% select(-sex)
  f_values - m_values
}) %>%
  mutate(iteration=c(1:1000)) %>% 
  pivot_longer(!iteration,names_to="trait",values_to="value") %>%
      group_by(trait) %>%
      summarise(mean=mean(value),
        p1 = quantile(value, probs = 0.1),
        p9 = quantile(value, probs = 0.9),
        positive_support = sum(value > 0)/1000,
        negative_support = sum(value < 0)/1000) %>%
      arrange(-positive_support)


```

```{r element_sex_plot_red, message=FALSE, warning=FALSE, fig.dim=c(10, 8)}

# male associated
element_pred_sex_red %>%
  filter(mean >0, positive_support>=0.9) %>%
  arrange(-mean) %>%
  select(-negative_support) %>%
  rename(support=positive_support) %>%
  left_join(GIFT_db2 %>% select(Code_element, Element, Function),by=join_by(trait==Code_element)) %>%
  relocate(Element, Function, .after=trait) %>%
  unique() %>%
  paged_table()

# female associated
element_pred_sex_red %>%
  filter(mean <0, negative_support>=0.9) %>%
  arrange(mean) %>%
  select(-positive_support) %>%
  rename(support=negative_support) %>%
  left_join(GIFT_db2 %>% select(Code_element, Element, Function),by=join_by(trait==Code_element)) %>%
  relocate(Element, Function, .after=trait) %>%
  unique() %>%
  paged_table()

positive_red <- element_pred_sex_red%>%
  filter(mean >0) %>%
  arrange(mean) %>%
  filter(positive_support>=0.9) %>%
  select(-negative_support) %>%
  rename(support=positive_support)
negative_red <- element_pred_sex_red %>%
  filter(mean <0) %>%
  arrange(mean) %>%
  filter(negative_support>=0.9) %>%
  select(-positive_support) %>%
  rename(support=negative_support)

all_elements_red <- bind_rows(positive_red,negative_red ) %>%
  left_join(GIFT_db2,by=join_by(trait==Code_element)) %>%
  mutate(trait=factor(trait,levels=c(rev(positive_red$trait),rev(negative_red$trait)))) %>%
  mutate(Code_function=factor(Code_function)) %>%
  mutate(element_legend=str_c(trait," - ",Element)) %>%
  mutate(function_legend=str_c(Code_function," - ",Function)) %>%
  select(trait,mean,p1,p9,element_legend,function_legend) %>% 
  unique()

red_elem_sex_plot <- all_elements_red %>%
  ggplot(aes(x=mean, y=fct_reorder(element_legend, mean), xmin=p1, xmax=p9, color=function_legend)) +
      geom_point() +
      geom_errorbar() +
      xlim(c(-3.5,3.5)) +
      geom_vline(xintercept=0) +
      scale_color_manual(values = gift_colors$Color) +
      theme_minimal() +
      labs(x="Difference in log-abundance",y="Functional trait", title ="S. carolinensis", color="Function") +
      annotate('text', x=-2, y=10, label = "Enriched in\nfemales", color='black') + #NB reversed
      annotate('text', x=2, y=10, label = "Enriched in\nmales", color='black') +
      #guides(col = guide_legend(ncol = 1))
      theme(legend.position = "right",
            axis.title.y = element_blank(),
            plot.title=element_text( hjust=0.5, vjust=0.6, face='bold', size=12))


```

```{r element_sex_pred_grey, results="hide", eval=TRUE, message=FALSE, warning=FALSE, cache=TRUE}

community_elem_sex_grey <- pred_sex_grey  %>%
  filter(genome %in% pred_mags.red) %>% #keep only predictive mags
  group_by(sex, genome) %>%
  mutate(row_id = row_number()) %>%
  pivot_wider(names_from = genome, values_from = value) %>%
  ungroup() %>%
  group_split(row_id) %>%
  as.list() %>%
  lapply(., FUN = function(x){x %>%
    select(-row_id) %>%
    column_to_rownames(var = "sex") %>%
    as.data.frame() %>%
    exp() %>%
    t() %>%
    tss() %>%
    to.community(GIFTs_elements,.,GIFT_db2) %>%
    as.data.frame() %>%
    rownames_to_column(var="sex")
   })

element_pred_sex_grey <- map_dfr(community_elem_sex_grey, function(df) {
  f_values <- df %>% filter(sex == "female") %>% select(-sex)
  m_values <- df %>% filter(sex == "male") %>% select(-sex)
  f_values - m_values
}) %>%
  mutate(iteration=c(1:1000)) %>% 
  pivot_longer(!iteration,names_to="trait",values_to="value") %>%
      group_by(trait) %>%
      summarise(mean=mean(value),
        p1 = quantile(value, probs = 0.1),
        p9 = quantile(value, probs = 0.9),
        positive_support = sum(value > 0)/1000,
        negative_support = sum(value < 0)/1000) %>%
      arrange(-positive_support)


```

```{r element_sex_plot_grey, message=FALSE, warning=FALSE, fig.dim=c(10, 8)}

# male associated
element_pred_sex_grey %>%
  filter(mean >0, positive_support>=0.9) %>%
  arrange(-mean) %>%
  select(-negative_support) %>%
  rename(support=positive_support) %>%
  left_join(GIFT_db2 %>% select(Code_element, Element, Function),by=join_by(trait==Code_element)) %>%
  relocate(Element, Function, .after=trait) %>%
  unique() %>%
  paged_table()

# female associated
element_pred_sex_grey %>%
  filter(mean <0, negative_support>=0.9) %>%
  arrange(mean) %>%
  select(-positive_support) %>%
  rename(support=negative_support) %>%
  left_join(GIFT_db2 %>% select(Code_element, Element, Function),by=join_by(trait==Code_element)) %>%
  relocate(Element, Function, .after=trait) %>%
  unique() %>%
  paged_table()

positive_grey <- element_pred_sex_grey%>%
  filter(mean >0) %>%
  arrange(mean) %>%
  filter(positive_support>=0.9) %>%
  select(-negative_support) %>%
  rename(support=positive_support)
negative_grey <- element_pred_sex_grey %>%
  filter(mean <0) %>%
  arrange(mean) %>%
  filter(negative_support>=0.9) %>%
  select(-positive_support) %>%
  rename(support=negative_support)

all_elements_grey <- bind_rows(positive_grey,negative_grey ) %>%
  left_join(GIFT_db2,by=join_by(trait==Code_element)) %>%
  mutate(trait=factor(trait,levels=c(rev(positive_grey$trait),rev(negative_grey$trait)))) %>%
  mutate(Code_function=factor(Code_function)) %>%
  mutate(element_legend=str_c(trait," - ",Element)) %>%
  mutate(function_legend=str_c(Code_function," - ",Function)) %>%
  select(trait,mean,p1,p9,element_legend,function_legend) %>% 
  unique()

grey_elem_sex_plot <- all_elements_grey %>%
  ggplot(aes(x=mean, y=fct_reorder(element_legend, mean), xmin=p1, xmax=p9, color=function_legend)) +
      geom_point() +
      geom_errorbar() +
      xlim(c(-3.5,3.5)) +
      geom_vline(xintercept=0) +
      scale_color_manual(values = gift_colors$Color) +
      theme_minimal() +
      labs(x="Difference in log-abundance",y="Functional trait", title ="S. carolinensis", color="Function") +
      annotate('text', x=-2, y=2, label = "Enriched in\nfemales", color='black') + #NB reversed
      annotate('text', x=2, y=2, label = "Enriched in\nmales", color='black') +
      #guides(col = guide_legend(ncol = 1))
      theme(legend.position = "right",
            axis.title.y = element_blank(),
            plot.title=element_text( hjust=0.5, vjust=0.6, face='bold', size=12))

```

```{r elem_sex_plot_all, fig.dim=c(10,10)}

ggarrange(red_elem_sex_plot, grey_elem_sex_plot, legend="right", common.legend = TRUE,
          #labels = c("A", "B", "C"),
          ncol = 1, nrow = 2, align="v")

```
